+++ kernel-module-imx-gpu-viv-src/hal/kernel/arch/gc_hal_kernel_hardware.c
@@ -2829,7 +2829,7 @@ gckHARDWARE_InitializeHardware(
             gckOS_WriteRegisterEx(Hardware->os,
                                   Hardware->core,
                                   0x0010C,
-                                  ((((gctUINT32) (0x01590880)) & ~(((gctUINT32) (((gctUINT32) ((((1 ?
+                                  ((((gctUINT32) (0x015B0880)) & ~(((gctUINT32) (((gctUINT32) ((((1 ?
  23:23) - (0 ?
  23:23) + 1) == 32) ?
  ~0U : (~(~0U << ((1 ?
@@ -2851,7 +2851,7 @@ gckHARDWARE_InitializeHardware(
     {
         gctUINT32 pulseEater;
 
-        pulseEater = ((((gctUINT32) (0x01590880)) & ~(((gctUINT32) (((gctUINT32) ((((1 ?
+        pulseEater = ((((gctUINT32) (0x015B0880)) & ~(((gctUINT32) (((gctUINT32) ((((1 ?
  16:16) - (0 ?
  16:16) + 1) == 32) ?
  ~0U : (~(~0U << ((1 ?
@@ -10799,17 +10799,17 @@ OnError:
 
 static gctUINT32
 CalcDelta(
-    IN gctUINT32 new,
-    IN gctUINT32 old
+    IN gctUINT32 newval,
+    IN gctUINT32 oldval
     )
 {
-    if (new >= old)
+    if (newval >= oldval)
     {
-        return new - old;
+        return newval - oldval;
     }
     else
     {
-        return (gctUINT32)((gctUINT64)new + 0x100000000ll - old);
+        return (gctUINT32)((gctUINT64)newval + 0x100000000ll - oldval);
     }
 }
 
@@ -12967,7 +12967,7 @@ _ResetGPU(
                     0x00104,
                     0x00000000));
 
-        control = ((((gctUINT32) (0x01590880)) & ~(((gctUINT32) (((gctUINT32) ((((1 ?
+        control = ((((gctUINT32) (0x015B0880)) & ~(((gctUINT32) (((gctUINT32) ((((1 ?
  17:17) - (0 ?
  17:17) + 1) == 32) ?
  ~0U : (~(~0U << ((1 ?
