+++ kernel-module-imx-gpu-viv-src/hal/kernel/gc_hal_kernel.h
@@ -2362,6 +2362,14 @@ gckDEVICE_Dispatch(
     IN gcsHAL_INTERFACE_PTR Interface
     );
 
+#if VIVANTE_PROFILER
+gceSTATUS
+gckDEVICE_Profiler_Dispatch(
+    IN gckDEVICE Device,
+    IN gcsHAL_PROFILER_INTERFACE_PTR Interface
+    );
+#endif
+
 gceSTATUS
 gckDEVICE_GetMMU(
     IN gckDEVICE Device,
