+++ kernel-module-imx-gpu-viv-src/hal/kernel/inc/gc_hal_base.h
@@ -59,6 +59,7 @@
 #include "gc_hal_enum.h"
 #include "gc_hal_types.h"
 #include "gc_hal_debug_zones.h"
+#include "shared/gc_hal_base_shared.h"
 
 
 #ifdef __cplusplus
@@ -102,14 +103,6 @@ typedef void *                          gcoVG;
 
 typedef struct _gcoFENCE *              gcoFENCE;
 typedef struct _gcsSYNC_CONTEXT *       gcsSYNC_CONTEXT_PTR;
-
-typedef enum {
-    gcvFENCE_TYPE_READ          = 0x1,
-    gcvFENCE_TYPE_WRITE         = 0x2,
-    gcvFENCE_TYPE_ALL           = gcvFENCE_TYPE_READ | gcvFENCE_TYPE_WRITE,
-    gcvFNECE_TYPE_INVALID       = 0x10000,
-}
-gceFENCE_TYPE;
 
 typedef struct _gcsUSER_MEMORY_DESC *   gcsUSER_MEMORY_DESC_PTR;
 
@@ -305,19 +298,6 @@ typedef struct _gcsDRIVER_TLS
 }
 gcsDRIVER_TLS;
 
-typedef enum _gceTLS_KEY
-{
-    gcvTLS_KEY_EGL,
-    gcvTLS_KEY_OPENGL_ES,
-    gcvTLS_KEY_OPENVG,
-    gcvTLS_KEY_OPENGL,
-    gcvTLS_KEY_OPENCL,
-    gcvTLS_KEY_OPENVX,
-
-    gcvTLS_KEY_COUNT
-}
-gceTLS_KEY;
-
 typedef struct _gcsTLS * gcsTLS_PTR;
 
 typedef struct _gcsTLS
@@ -367,139 +347,6 @@ typedef struct _gcsTLS
 }
 gcsTLS;
 
-/******************************************************************************\
-********************************* Enumerations *********************************
-\******************************************************************************/
-
-typedef enum _gcePLS_VALUE
-{
-  gcePLS_VALUE_EGL_DISPLAY_INFO,
-  gcePLS_VALUE_EGL_CONFIG_FORMAT_INFO,
-  gcePLS_VALUE_EGL_DESTRUCTOR_INFO,
-}
-gcePLS_VALUE;
-
-/* Video memory pool type. */
-typedef enum _gcePOOL
-{
-    gcvPOOL_UNKNOWN = 0,
-    gcvPOOL_DEFAULT,
-    gcvPOOL_LOCAL,
-    gcvPOOL_LOCAL_INTERNAL,
-    gcvPOOL_LOCAL_EXTERNAL,
-    gcvPOOL_UNIFIED,
-    gcvPOOL_SYSTEM,
-    gcvPOOL_SRAM,
-    gcvPOOL_VIRTUAL,
-    gcvPOOL_USER,
-    gcvPOOL_INTERNAL_SRAM,
-    gcvPOOL_EXTERNAL_SRAM,
-
-    gcvPOOL_NUMBER_OF_POOLS
-}
-gcePOOL;
-
-#if gcdENABLE_3D
-/* Blending functions. */
-typedef enum _gceBLEND_FUNCTION
-{
-    gcvBLEND_ZERO,
-    gcvBLEND_ONE,
-    gcvBLEND_SOURCE_COLOR,
-    gcvBLEND_INV_SOURCE_COLOR,
-    gcvBLEND_SOURCE_ALPHA,
-    gcvBLEND_INV_SOURCE_ALPHA,
-    gcvBLEND_TARGET_COLOR,
-    gcvBLEND_INV_TARGET_COLOR,
-    gcvBLEND_TARGET_ALPHA,
-    gcvBLEND_INV_TARGET_ALPHA,
-    gcvBLEND_SOURCE_ALPHA_SATURATE,
-    gcvBLEND_CONST_COLOR,
-    gcvBLEND_INV_CONST_COLOR,
-    gcvBLEND_CONST_ALPHA,
-    gcvBLEND_INV_CONST_ALPHA,
-}
-gceBLEND_FUNCTION;
-
-/* Blending modes. */
-typedef enum _gceBLEND_MODE
-{
-    gcvBLEND_ADD = 0,
-    gcvBLEND_SUBTRACT,
-    gcvBLEND_REVERSE_SUBTRACT,
-    gcvBLEND_MIN,
-    gcvBLEND_MAX,
-    gcvBLEND_MULTIPLY,
-    gcvBLEND_SCREEN,
-    gcvBLEND_OVERLAY,
-    gcvBLEND_DARKEN,
-    gcvBLEND_LIGHTEN,
-    gcvBLEND_COLORDODGE,
-    gcvBLEND_COLORBURN,
-    gcvBLEND_HARDLIGHT,
-    gcvBLEND_SOFTLIGHT,
-    gcvBLEND_DIFFERENCE,
-    gcvBLEND_EXCLUSION,
-    gcvBLEND_HSL_HUE,
-    gcvBLEND_HSL_SATURATION,
-    gcvBLEND_HSL_COLOR,
-    gcvBLEND_HSL_LUMINOSITY,
-
-    gcvBLEND_TOTAL
-}
-gceBLEND_MODE;
-
-/* Depth modes. */
-typedef enum _gceDEPTH_MODE
-{
-    gcvDEPTH_NONE,
-    gcvDEPTH_Z,
-    gcvDEPTH_W,
-}
-gceDEPTH_MODE;
-#endif /* gcdENABLE_3D */
-
-
-/* API flags. */
-typedef enum _gceAPI
-{
-    gcvAPI_D3D = 1,
-    gcvAPI_OPENGL_ES11,
-    gcvAPI_OPENGL_ES20,
-    gcvAPI_OPENGL_ES30,
-    gcvAPI_OPENGL_ES31,
-    gcvAPI_OPENGL_ES32,
-    gcvAPI_OPENGL,
-    gcvAPI_OPENVG,
-    gcvAPI_OPENCL,
-    gcvAPI_OPENVK,
-}
-gceAPI;
-
-typedef enum _gceWHERE
-{
-    gcvWHERE_COMMAND_PREFETCH = 0,
-    gcvWHERE_COMMAND,
-    gcvWHERE_RASTER,
-    gcvWHERE_PIXEL,
-    gcvWHERE_BLT,
-}
-gceWHERE;
-
-typedef enum _gceHOW
-{
-    gcvHOW_SEMAPHORE            = 0x1,
-    gcvHOW_STALL                = 0x2,
-    gcvHOW_SEMAPHORE_STALL      = 0x3,
-}
-gceHOW;
-
-typedef enum _gceSignalHandlerType
-{
-    gcvHANDLE_SIGFPE_WHEN_SIGNAL_CODE_IS_0        = 0x1,
-}
-gceSignalHandlerType;
-
 typedef struct _gcsSURF_VIEW
 {
     gcoSURF surf;
@@ -536,20 +383,6 @@ typedef struct _gcsHAL_CHIPIDENTITY
     gctUINT64                   platformFlagBits;
 }
 gcsHAL_CHIPIDENTITY;
-
-
-#define gcdEXTERNAL_MEMORY_NAME_MAX 32
-#define gcdEXTERNAL_MEMORY_DATA_MAX 8
-
-typedef struct _gcsEXTERNAL_MEMORY_INFO
-{
-    /* Name of allocator used to attach this memory. */
-    gctCHAR                allocatorName[gcdEXTERNAL_MEMORY_NAME_MAX];
-
-    /* User defined data which will be passed to allocator. */
-    gctUINT32              userData[gcdEXTERNAL_MEMORY_DATA_MAX];
-}
-gcsEXTERNAL_MEMORY_INFO;
 
 /******************************************************************************\
 ********************************* gcoHAL Object *********************************
@@ -1393,17 +1226,6 @@ gcoOS_DeviceControl(
 
 #define gcdMAX_PATH 512
 
-typedef enum _gceFILE_MODE
-{
-    gcvFILE_CREATE          = 0,
-    gcvFILE_APPEND,
-    gcvFILE_READ,
-    gcvFILE_CREATETEXT,
-    gcvFILE_APPENDTEXT,
-    gcvFILE_READTEXT,
-}
-gceFILE_MODE;
-
 /* Open a file. */
 gceSTATUS
 gcoOS_Open(
@@ -1575,14 +1397,6 @@ gcoOS_Stat(
     IN gctCONST_STRING FileName,
     OUT gctPOINTER Buffer
     );
-
-typedef enum _gceFILE_WHENCE
-{
-    gcvFILE_SEEK_SET,
-    gcvFILE_SEEK_CUR,
-    gcvFILE_SEEK_END
-}
-gceFILE_WHENCE;
 
 /* Set the current position of a file. */
 gceSTATUS
@@ -2284,48 +2098,6 @@ typedef struct _gcsPIXEL
 /*----------------------------------------------------------------------------*/
 /*------------------------------- gcoSURF Common ------------------------------*/
 
-/* Color format classes. */
-typedef enum _gceFORMAT_CLASS
-{
-    gcvFORMAT_CLASS_RGBA        = 4500,
-    gcvFORMAT_CLASS_YUV,
-    gcvFORMAT_CLASS_INDEX,
-    gcvFORMAT_CLASS_LUMINANCE,
-    gcvFORMAT_CLASS_BUMP,
-    gcvFORMAT_CLASS_DEPTH,
-    gcvFORMAT_CLASS_ASTC,
-    gcvFORMAT_CLASS_COMPRESSED,
-    gcvFORMAT_CLASS_OTHER
-}
-gceFORMAT_CLASS;
-
-/* Color format data type */
-typedef enum _gceFORMAT_DATATYPE
-{
-    gcvFORMAT_DATATYPE_UNSIGNED_NORMALIZED,
-    gcvFORMAT_DATATYPE_SIGNED_NORMALIZED,
-    gcvFORMAT_DATATYPE_UNSIGNED_INTEGER,
-    gcvFORMAT_DATATYPE_SIGNED_INTEGER,
-    gcvFORMAT_DATATYPE_FLOAT16,
-    gcvFORMAT_DATATYPE_FLOAT32,
-    gcvFORMAT_DATATYPE_FLOAT_E5B9G9R9,
-    gcvFORMAT_DATATYPE_FLOAT_B10G11R11F,
-    gcvFORMAT_DATATYPE_INDEX,
-    gcvFORMAT_DATATYPE_SRGB,
-    gcvFORMAT_DATATYPE_FLOAT32_UINT,
-}
-gceFORMAT_DATATYPE;
-
-/* Special enums for width field in gcsFORMAT_COMPONENT. */
-typedef enum _gceCOMPONENT_CONTROL
-{
-    gcvCOMPONENT_NOTPRESENT     = 0x00,
-    gcvCOMPONENT_DONTCARE       = 0x80,
-    gcvCOMPONENT_WIDTHMASK      = 0x7F,
-    gcvCOMPONENT_ODD            = 0x80
-}
-gceCOMPONENT_CONTROL;
-
 /* Color format component parameters. */
 typedef struct _gcsFORMAT_COMPONENT
 {
@@ -2477,14 +2249,6 @@ extern gcsFORMAT_COMPONENT gcvPIXEL_COMP_XXX8;
 extern gcsFORMAT_COMPONENT gcvPIXEL_COMP_XX8X;
 extern gcsFORMAT_COMPONENT gcvPIXEL_COMP_X8XX;
 extern gcsFORMAT_COMPONENT gcvPIXEL_COMP_8XXX;
-
-typedef enum _gceORIENTATION
-{
-    gcvORIENTATION_TOP_BOTTOM,
-    gcvORIENTATION_BOTTOM_TOP,
-}
-gceORIENTATION;
-
 
 /* Construct a new gcoSURF object. */
 gceSTATUS
@@ -3552,18 +3316,6 @@ gcoOS_StackRemove(
 **      Text        Optional text
 **      ...         Optional arguments for text.
 */
-typedef struct _gcsBINARY_TRACE_MESSAGE * gcsBINARY_TRACE_MESSAGE_PTR;
-typedef struct _gcsBINARY_TRACE_MESSAGE
-{
-    gctUINT32   signature;
-    gctUINT32   pid;
-    gctUINT32   tid;
-    gctUINT32   line;
-    gctUINT32   numArguments;
-    gctUINT8    payload;
-}
-gcsBINARY_TRACE_MESSAGE;
-
 void
 gcoOS_BinaryTrace(
     IN gctCONST_STRING Function,
@@ -3579,8 +3331,6 @@ gckOS_BinaryTrace(
     IN gctCONST_STRING Text OPTIONAL,
     ...
     );
-
-#define gcdBINARY_TRACE_MESSAGE_SIZE 240
 
 #if gcdBINARY_TRACE
 #   define gcmBINARY_TRACE          gcoOS_BinaryTrace
@@ -3974,33 +3724,6 @@ gcoOS_Print(
 #   define gcmPRINT_VERSION()       do { gcmSTACK_DUMP(); } while (gcvFALSE)
 #   define gcmkPRINT_VERSION()      do { } while (gcvFALSE)
 #endif
-
-typedef enum _gceDUMP_BUFFER_TYPE
-{
-    gcvDUMP_BUFFER_USER_STRING,
-    gcvDUMP_BUFFER_VERIFY,
-
-    gcvDUMP_BUFFER_MEMORY,
-    gcvDUMP_BUFFER_TEXTURE,
-    gcvDUMP_BUFFER_STREAM,
-    gcvDUMP_BUFFER_INDEX,
-    gcvDUMP_BUFFER_BUFOBJ,
-    gcvDUMP_BUFFER_IMAGE,
-    /* A type of command, but should not execute directly. */
-    gcvDUMP_BUFFER_INSTRUCTION,
-    gcvDUMP_BUFFER_CONTEXT,
-    gcvDUMP_BUFFER_COMMAND,
-    gcvDUMP_BUFFER_ASYNC_COMMAND,
-    gcvDUMP_BUFFER_USER_TYPE_LAST = gcvDUMP_BUFFER_ASYNC_COMMAND,
-
-    gcvDUMP_BUFFER_KERNEL_CONTEXT,
-    gcvDUMP_BUFFER_KERNEL_COMMAND,
-
-    gcvDUMP_BUFFER_PHYSICAL_MEMORY,
-
-    gcvDUMP_BUFFER_TYPE_COUNT,
-}
-gceDUMP_BUFFER_TYPE;
 
 void
 gckOS_Dump(
@@ -4923,15 +4646,6 @@ gckOS_DebugStatus2Name(
 /******************************************************************************\
 ****************************** User Debug Option ******************************
 \******************************************************************************/
-
-/* User option. */
-typedef enum _gceDEBUG_MSG
-{
-    gcvDEBUG_MSG_NONE,
-    gcvDEBUG_MSG_ERROR,
-    gcvDEBUG_MSG_WARNING
-}
-gceDEBUG_MSG;
 
 typedef struct _gcsUSER_DEBUG_OPTION
 {
@@ -6048,7 +5762,6 @@ typedef struct _memory_profile_info
         gctUINT32     total_freeCount;
     } system_memory, gpu_memory;
 } memory_profile_info;
-
 
 gceSTATUS
 gcoOS_GetMemoryProfileInfo(
