+++ kernel-module-imx-gpu-viv-src/hal/kernel/inc/gc_hal_engine_vg.h
@@ -63,316 +63,6 @@ extern "C" {
 #include "gc_hal_types.h"
 
 /******************************************************************************\
-******************************** VG Enumerations *******************************
-\******************************************************************************/
-
-/**
-**  @ingroup gcoVG
-**
-**  @brief  Tiling mode for painting and imagig.
-**
-**  This enumeration defines the tiling modes supported by the HAL.  This is
-**  in fact a one-to-one mapping of the OpenVG 1.1 tile modes.
-*/
-typedef enum _gceTILE_MODE
-{
-    gcvTILE_FILL,
-    gcvTILE_PAD,
-    gcvTILE_REPEAT,
-    gcvTILE_REFLECT
-}
-gceTILE_MODE;
-
-/******************************************************************************/
-/** @ingroup gcoVG
-**
-**  @brief  The different paint modes.
-**
-**  This enumeration lists the available paint modes.
-*/
-typedef enum _gcePAINT_TYPE
-{
-    /** Solid color. */
-    gcvPAINT_MODE_SOLID,
-
-    /** Linear gradient. */
-    gcvPAINT_MODE_LINEAR,
-
-    /** Radial gradient. */
-    gcvPAINT_MODE_RADIAL,
-
-    /** Pattern. */
-    gcvPAINT_MODE_PATTERN,
-
-    /** Mode count. */
-    gcvPAINT_MODE_COUNT
-}
-gcePAINT_TYPE;
-
-/**
-** @ingroup gcoVG
-**
-**  @brief Types of path data supported by HAL.
-**
-**  This enumeration defines the types of path data supported by the HAL.
-**  This is in fact a one-to-one mapping of the OpenVG 1.1 path types.
-*/
-typedef enum _gcePATHTYPE
-{
-    gcePATHTYPE_UNKNOWN = -1,
-    gcePATHTYPE_INT8,
-    gcePATHTYPE_INT16,
-    gcePATHTYPE_INT32,
-    gcePATHTYPE_FLOAT
-}
-gcePATHTYPE;
-
-/**
-** @ingroup gcoVG
-**
-**  @brief Supported path segment commands.
-**
-**  This enumeration defines the path segment commands supported by the HAL.
-*/
-typedef enum _gceVGCMD
-{
-    gcvVGCMD_END, /*  0: 0x00           */
-    gcvVGCMD_CLOSE, /*  1: 0x01         */
-    gcvVGCMD_MOVE, /*  2: 0x02          */
-    gcvVGCMD_MOVE_REL, /*  3: 0x03      */
-    gcvVGCMD_LINE, /*  4: 0x04          */
-    gcvVGCMD_LINE_REL, /*  5: 0x05      */
-    gcvVGCMD_QUAD, /*  6: 0x06     */
-    gcvVGCMD_QUAD_REL, /*  7: 0x07 */
-    gcvVGCMD_CUBIC, /*  8: 0x08         */
-    gcvVGCMD_CUBIC_REL, /*  9: 0x09     */
-    gcvVGCMD_BREAK, /* 10: 0x0A         */
-    gcvVGCMD_HLINE, /* 11: ******* R E S E R V E D *******/
-    gcvVGCMD_HLINE_REL, /* 12: ******* R E S E R V E D *******/
-    gcvVGCMD_VLINE, /* 13: ******* R E S E R V E D *******/
-    gcvVGCMD_VLINE_REL, /* 14: ******* R E S E R V E D *******/
-    gcvVGCMD_SQUAD, /* 15: ******* R E S E R V E D *******/
-    gcvVGCMD_SQUAD_REL, /* 16: ******* R E S E R V E D *******/
-    gcvVGCMD_SCUBIC, /* 17: ******* R E S E R V E D *******/
-    gcvVGCMD_SCUBIC_REL, /* 18: ******* R E S E R V E D *******/
-    gcvVGCMD_SCCWARC, /* 19: ******* R E S E R V E D *******/
-    gcvVGCMD_SCCWARC_REL, /* 20: ******* R E S E R V E D *******/
-    gcvVGCMD_SCWARC, /* 21: ******* R E S E R V E D *******/
-    gcvVGCMD_SCWARC_REL, /* 22: ******* R E S E R V E D *******/
-    gcvVGCMD_LCCWARC, /* 23: ******* R E S E R V E D *******/
-    gcvVGCMD_LCCWARC_REL, /* 24: ******* R E S E R V E D *******/
-    gcvVGCMD_LCWARC, /* 25: ******* R E S E R V E D *******/
-    gcvVGCMD_LCWARC_REL, /* 26: ******* R E S E R V E D *******/
-
-    /* The width of the command recognized by the hardware on bits. */
-    gcvVGCMD_WIDTH = 5,
-
-    /* Hardware command mask. */
-    gcvVGCMD_MASK = (1 << gcvVGCMD_WIDTH) - 1,
-
-    /* Command modifiers. */
-    gcvVGCMD_H_MOD   = 1 << gcvVGCMD_WIDTH, /* =  32 */
-    gcvVGCMD_V_MOD   = 2 << gcvVGCMD_WIDTH, /* =  64 */
-    gcvVGCMD_S_MOD   = 3 << gcvVGCMD_WIDTH, /* =  96 */
-    gcvVGCMD_ARC_MOD = 4 << gcvVGCMD_WIDTH, /* = 128 */
-
-    /* Emulated LINE commands. */
-    gcvVGCMD_HLINE_EMUL     = gcvVGCMD_H_MOD | gcvVGCMD_LINE, /* =  36 */
-    gcvVGCMD_HLINE_EMUL_REL = gcvVGCMD_H_MOD | gcvVGCMD_LINE_REL, /* =  37 */
-    gcvVGCMD_VLINE_EMUL     = gcvVGCMD_V_MOD | gcvVGCMD_LINE, /* =  68 */
-    gcvVGCMD_VLINE_EMUL_REL = gcvVGCMD_V_MOD | gcvVGCMD_LINE_REL, /* =  69 */
-
-    /* Emulated SMOOTH commands. */
-    gcvVGCMD_SQUAD_EMUL      = gcvVGCMD_S_MOD | gcvVGCMD_QUAD, /* = 102 */
-    gcvVGCMD_SQUAD_EMUL_REL  = gcvVGCMD_S_MOD | gcvVGCMD_QUAD_REL, /* = 103 */
-    gcvVGCMD_SCUBIC_EMUL     = gcvVGCMD_S_MOD | gcvVGCMD_CUBIC, /* = 104 */
-    gcvVGCMD_SCUBIC_EMUL_REL = gcvVGCMD_S_MOD | gcvVGCMD_CUBIC_REL, /* = 105 */
-
-    /* Emulation ARC commands. */
-    gcvVGCMD_ARC_LINE     = gcvVGCMD_ARC_MOD | gcvVGCMD_LINE, /* = 132 */
-    gcvVGCMD_ARC_LINE_REL = gcvVGCMD_ARC_MOD | gcvVGCMD_LINE_REL, /* = 133 */
-    gcvVGCMD_ARC_QUAD     = gcvVGCMD_ARC_MOD | gcvVGCMD_QUAD, /* = 134 */
-    gcvVGCMD_ARC_QUAD_REL = gcvVGCMD_ARC_MOD | gcvVGCMD_QUAD_REL     /* = 135 */
-}
-gceVGCMD;
-typedef enum _gceVGCMD * gceVGCMD_PTR;
-
-/**
-**  @ingroup gcoVG
-**
-**  @brief  Blending modes supported by the HAL.
-**
-**  This enumeration defines the blending modes supported by the HAL.  This is
-**  in fact a one-to-one mapping of the OpenVG 1.1 blending modes.
-*/
-typedef enum _gceVG_BLEND
-{
-    gcvVG_BLEND_SRC,
-    gcvVG_BLEND_SRC_OVER,
-    gcvVG_BLEND_DST_OVER,
-    gcvVG_BLEND_SRC_IN,
-    gcvVG_BLEND_DST_IN,
-    gcvVG_BLEND_MULTIPLY,
-    gcvVG_BLEND_SCREEN,
-    gcvVG_BLEND_DARKEN,
-    gcvVG_BLEND_LIGHTEN,
-    gcvVG_BLEND_ADDITIVE,
-    gcvVG_BLEND_SUBTRACT,
-    gcvVG_BLEND_FILTER
-}
-gceVG_BLEND;
-
-/**
-**  @ingroup gcoVG
-**
-**  @brief  Image modes supported by the HAL.
-**
-**  This enumeration defines the image modes supported by the HAL.  This is
-**  in fact a one-to-one mapping of the OpenVG 1.1 image modes with the addition
-**  of NO IMAGE.
-*/
-typedef enum _gceVG_IMAGE
-{
-    gcvVG_IMAGE_NONE,
-    gcvVG_IMAGE_NORMAL,
-    gcvVG_IMAGE_MULTIPLY,
-    gcvVG_IMAGE_STENCIL,
-    gcvVG_IMAGE_FILTER
-}
-gceVG_IMAGE;
-
-/**
-**  @ingroup gcoVG
-**
-**  @brief  Filter mode patterns and imaging.
-**
-**  This enumeration defines the filter modes supported by the HAL.
-*/
-typedef enum _gceIMAGE_FILTER
-{
-    gcvFILTER_POINT,
-    gcvFILTER_LINEAR,
-    gcvFILTER_BI_LINEAR
-}
-gceIMAGE_FILTER;
-
-/**
-**  @ingroup gcoVG
-**
-**  @brief  Primitive modes supported by the HAL.
-**
-**  This enumeration defines the primitive modes supported by the HAL.
-*/
-typedef enum _gceVG_PRIMITIVE
-{
-    gcvVG_SCANLINE,
-    gcvVG_RECTANGLE,
-    gcvVG_TESSELLATED,
-    gcvVG_TESSELLATED_TILED
-}
-gceVG_PRIMITIVE;
-
-/**
-**  @ingroup gcoVG
-**
-**  @brief  Rendering quality modes supported by the HAL.
-**
-**  This enumeration defines the rendering quality modes supported by the HAL.
-*/
-typedef enum _gceRENDER_QUALITY
-{
-    gcvVG_NONANTIALIASED,
-    gcvVG_2X2_MSAA,
-    gcvVG_2X4_MSAA,
-    gcvVG_4X4_MSAA
-}
-gceRENDER_QUALITY;
-
-/**
-**  @ingroup gcoVG
-**
-**  @brief  Fill rules supported by the HAL.
-**
-**  This enumeration defines the fill rules supported by the HAL.
-*/
-typedef enum _gceFILL_RULE
-{
-    gcvVG_EVEN_ODD,
-    gcvVG_NON_ZERO
-}
-gceFILL_RULE;
-
-/**
-**  @ingroup gcoVG
-**
-**  @brief  Cap styles supported by the HAL.
-**
-**  This enumeration defines the cap styles supported by the HAL.
-*/
-typedef enum _gceCAP_STYLE
-{
-    gcvCAP_BUTT,
-    gcvCAP_ROUND,
-    gcvCAP_SQUARE
-}
-gceCAP_STYLE;
-
-/**
-**  @ingroup gcoVG
-**
-**  @brief  Join styles supported by the HAL.
-**
-**  This enumeration defines the join styles supported by the HAL.
-*/
-typedef enum _gceJOIN_STYLE
-{
-    gcvJOIN_MITER,
-    gcvJOIN_ROUND,
-    gcvJOIN_BEVEL
-}
-gceJOIN_STYLE;
-
-/**
-**  @ingroup gcoVG
-**
-**  @brief  Channel mask values.
-**
-**  This enumeration defines the values for channel mask used in image
-**  filtering.
-*/
-
-/* Base values for channel mask definitions. */
-#define gcvCHANNEL_X    (0)
-#define gcvCHANNEL_R    (1 << 0)
-#define gcvCHANNEL_G    (1 << 1)
-#define gcvCHANNEL_B    (1 << 2)
-#define gcvCHANNEL_A    (1 << 3)
-
-typedef enum _gceCHANNEL
-{
-    gcvCHANNEL_XXXX = (gcvCHANNEL_X | gcvCHANNEL_X | gcvCHANNEL_X | gcvCHANNEL_X),
-    gcvCHANNEL_XXXA = (gcvCHANNEL_X | gcvCHANNEL_X | gcvCHANNEL_X | gcvCHANNEL_A),
-    gcvCHANNEL_XXBX = (gcvCHANNEL_X | gcvCHANNEL_X | gcvCHANNEL_B | gcvCHANNEL_X),
-    gcvCHANNEL_XXBA = (gcvCHANNEL_X | gcvCHANNEL_X | gcvCHANNEL_B | gcvCHANNEL_A),
-
-    gcvCHANNEL_XGXX = (gcvCHANNEL_X | gcvCHANNEL_G | gcvCHANNEL_X | gcvCHANNEL_X),
-    gcvCHANNEL_XGXA = (gcvCHANNEL_X | gcvCHANNEL_G | gcvCHANNEL_X | gcvCHANNEL_A),
-    gcvCHANNEL_XGBX = (gcvCHANNEL_X | gcvCHANNEL_G | gcvCHANNEL_B | gcvCHANNEL_X),
-    gcvCHANNEL_XGBA = (gcvCHANNEL_X | gcvCHANNEL_G | gcvCHANNEL_B | gcvCHANNEL_A),
-
-    gcvCHANNEL_RXXX = (gcvCHANNEL_R | gcvCHANNEL_X | gcvCHANNEL_X | gcvCHANNEL_X),
-    gcvCHANNEL_RXXA = (gcvCHANNEL_R | gcvCHANNEL_X | gcvCHANNEL_X | gcvCHANNEL_A),
-    gcvCHANNEL_RXBX = (gcvCHANNEL_R | gcvCHANNEL_X | gcvCHANNEL_B | gcvCHANNEL_X),
-    gcvCHANNEL_RXBA = (gcvCHANNEL_R | gcvCHANNEL_X | gcvCHANNEL_B | gcvCHANNEL_A),
-
-    gcvCHANNEL_RGXX = (gcvCHANNEL_R | gcvCHANNEL_G | gcvCHANNEL_X | gcvCHANNEL_X),
-    gcvCHANNEL_RGXA = (gcvCHANNEL_R | gcvCHANNEL_G | gcvCHANNEL_X | gcvCHANNEL_A),
-    gcvCHANNEL_RGBX = (gcvCHANNEL_R | gcvCHANNEL_G | gcvCHANNEL_B | gcvCHANNEL_X),
-    gcvCHANNEL_RGBA = (gcvCHANNEL_R | gcvCHANNEL_G | gcvCHANNEL_B | gcvCHANNEL_A),
-}
-gceCHANNEL;
-
-/******************************************************************************\
 ******************************** VG Structures *******************************
 \******************************************************************************/
