+++ kernel-module-imx-gpu-viv-src/hal/kernel/inc/gc_hal_enum.h
@@ -57,597 +57,11 @@
 #define __gc_hal_enum_h_
 
 #include "gc_hal_options.h"
+#include "shared/gc_hal_enum_shared.h"
 
 #ifdef __cplusplus
 extern "C" {
 #endif
-
-/* Chip models. */
-typedef enum _gceCHIPMODEL
-{
-    gcv200  = 0x0200,
-    gcv300  = 0x0300,
-    gcv320  = 0x0320,
-    gcv328  = 0x0328,
-    gcv350  = 0x0350,
-    gcv355  = 0x0355,
-    gcv400  = 0x0400,
-    gcv410  = 0x0410,
-    gcv420  = 0x0420,
-    gcv428  = 0x0428,
-    gcv450  = 0x0450,
-    gcv500  = 0x0500,
-    gcv520  = 0x0520,
-    gcv530  = 0x0530,
-    gcv600  = 0x0600,
-    gcv620  = 0x0620,
-    gcv700  = 0x0700,
-    gcv800  = 0x0800,
-    gcv860  = 0x0860,
-    gcv880  = 0x0880,
-    gcv900  = 0x0900,
-    gcv1000 = 0x1000,
-    gcv1500 = 0x1500,
-    gcv2000 = 0x2000,
-    gcv2100 = 0x2100,
-    gcv2200 = 0x2200,
-    gcv2500 = 0x2500,
-    gcv3000 = 0x3000,
-    gcv4000 = 0x4000,
-    gcv5000 = 0x5000,
-    gcv5200 = 0x5200,
-    gcv6400 = 0x6400,
-    gcv7000 = 0x7000,
-    gcv7400 = 0x7400,
-    gcv8000 = 0x8000,
-}
-gceCHIPMODEL;
-
-/* Chip features. */
-typedef enum _gceFEATURE
-{
-    gcvFEATURE_PIPE_2D = 0,
-    gcvFEATURE_PIPE_3D,
-    gcvFEATURE_PIPE_VG,
-    gcvFEATURE_DC,
-    gcvFEATURE_HIGH_DYNAMIC_RANGE,
-    gcvFEATURE_MODULE_CG,
-    gcvFEATURE_MIN_AREA,
-    gcvFEATURE_BUFFER_INTERLEAVING,
-    gcvFEATURE_BYTE_WRITE_2D,
-    gcvFEATURE_ENDIANNESS_CONFIG,
-    gcvFEATURE_DUAL_RETURN_BUS,
-    gcvFEATURE_DEBUG_MODE,
-    gcvFEATURE_YUY2_RENDER_TARGET,
-    gcvFEATURE_FRAGMENT_PROCESSOR,
-    gcvFEATURE_2DPE20,
-    gcvFEATURE_FAST_CLEAR,
-    gcvFEATURE_YUV420_TILER,
-    gcvFEATURE_YUY2_AVERAGING,
-    gcvFEATURE_FLIP_Y,
-    gcvFEATURE_EARLY_Z,
-    gcvFEATURE_COMPRESSION,
-    gcvFEATURE_MSAA,
-    gcvFEATURE_SPECIAL_ANTI_ALIASING,
-    gcvFEATURE_SPECIAL_MSAA_LOD,
-    gcvFEATURE_422_TEXTURE_COMPRESSION,
-    gcvFEATURE_DXT_TEXTURE_COMPRESSION,
-    gcvFEATURE_ETC1_TEXTURE_COMPRESSION,
-    gcvFEATURE_CORRECT_TEXTURE_CONVERTER,
-    gcvFEATURE_TEXTURE_8K,
-    gcvFEATURE_SCALER,
-    gcvFEATURE_YUV420_SCALER,
-    gcvFEATURE_SHADER_HAS_W,
-    gcvFEATURE_SHADER_HAS_SIGN,
-    gcvFEATURE_SHADER_HAS_FLOOR,
-    gcvFEATURE_SHADER_HAS_CEIL,
-    gcvFEATURE_SHADER_HAS_SQRT,
-    gcvFEATURE_SHADER_HAS_TRIG,
-    gcvFEATURE_HZ,
-    gcvFEATURE_CORRECT_STENCIL,
-    gcvFEATURE_VG20,
-    gcvFEATURE_VG_FILTER,
-    gcvFEATURE_VG21,
-    gcvFEATURE_VG_DOUBLE_BUFFER,
-    gcvFEATURE_VG_RESOLUTION_8K,
-    gcvFEATURE_MC20,
-    gcvFEATURE_SUPER_TILED,
-    gcvFEATURE_FAST_CLEAR_FLUSH,
-    gcvFEATURE_2D_FILTERBLIT_PLUS_ALPHABLEND,
-    gcvFEATURE_2D_DITHER,
-    gcvFEATURE_2D_A8_TARGET,
-    gcvFEATURE_2D_A8_NO_ALPHA,
-    gcvFEATURE_2D_FILTERBLIT_FULLROTATION,
-    gcvFEATURE_2D_BITBLIT_FULLROTATION,
-    gcvFEATURE_WIDE_LINE,
-    gcvFEATURE_FC_FLUSH_STALL,
-    gcvFEATURE_FULL_DIRECTFB,
-    gcvFEATURE_HALF_FLOAT_PIPE,
-    gcvFEATURE_LINE_LOOP,
-    gcvFEATURE_2D_YUV_BLIT,
-    gcvFEATURE_2D_TILING,
-    gcvFEATURE_NON_POWER_OF_TWO,
-    gcvFEATURE_3D_TEXTURE,
-    gcvFEATURE_TEXTURE_ARRAY,
-    gcvFEATURE_TILE_FILLER,
-    gcvFEATURE_LOGIC_OP,
-    gcvFEATURE_MIXED_STREAMS,
-    gcvFEATURE_2D_MULTI_SOURCE_BLT,
-    gcvFEATURE_END_EVENT,
-    gcvFEATURE_VERTEX_10_10_10_2,
-    gcvFEATURE_TEXTURE_10_10_10_2,
-    gcvFEATURE_TEXTURE_ANISOTROPIC_FILTERING,
-    gcvFEATURE_TEXTURE_FLOAT_HALF_FLOAT,
-    gcvFEATURE_2D_ROTATION_STALL_FIX,
-    gcvFEATURE_2D_MULTI_SOURCE_BLT_EX,
-    gcvFEATURE_BUG_FIXES10,
-    gcvFEATURE_2D_MINOR_TILING,
-    gcvFEATURE_TEX_COMPRRESSION_SUPERTILED, /* Supertiled compressed textures are supported. */
-    gcvFEATURE_FAST_MSAA,
-    gcvFEATURE_BUG_FIXED_INDEXED_TRIANGLE_STRIP,
-    gcvFEATURE_TEXTURE_TILE_STATUS_READ,
-    gcvFEATURE_DEPTH_BIAS_FIX,
-    gcvFEATURE_RECT_PRIMITIVE,
-    gcvFEATURE_BUG_FIXES11,
-    gcvFEATURE_SUPERTILED_TEXTURE,
-    gcvFEATURE_2D_NO_COLORBRUSH_INDEX8,
-    gcvFEATURE_RS_YUV_TARGET,
-    gcvFEATURE_2D_FC_SOURCE, /* For tilestatus compression feature*/
-    gcvFEATURE_2D_CC_NOAA_SOURCE,
-    gcvFEATURE_PE_DITHER_FIX,
-    gcvFEATURE_2D_YUV_SEPARATE_STRIDE,
-    gcvFEATURE_FRUSTUM_CLIP_FIX,
-    gcvFEATURE_TEXTURE_SWIZZLE,
-    gcvFEATURE_PRIMITIVE_RESTART,
-    gcvFEATURE_TEXTURE_LINEAR,
-    gcvFEATURE_TEXTURE_YUV_ASSEMBLER,
-    gcvFEATURE_LINEAR_RENDER_TARGET,
-    gcvFEATURE_SHADER_HAS_ATOMIC,
-    gcvFEATURE_SHADER_HAS_INSTRUCTION_CACHE,
-    gcvFEATURE_SHADER_ENHANCEMENTS2,
-    gcvFEATURE_BUG_FIXES7,
-    gcvFEATURE_SHADER_HAS_RTNE,
-    gcvFEATURE_SHADER_HAS_EXTRA_INSTRUCTIONS2,
-    gcvFEATURE_SHADER_ENHANCEMENTS3,
-    gcvFEATURE_DYNAMIC_FREQUENCY_SCALING,
-    gcvFEATURE_SINGLE_BUFFER,
-    gcvFEATURE_OCCLUSION_QUERY,
-    gcvFEATURE_2D_GAMMA,
-    gcvFEATURE_2D_COLOR_SPACE_CONVERSION,
-    gcvFEATURE_2D_SUPER_TILE_VERSION,
-    gcvFEATURE_HALTI0,
-    gcvFEATURE_HALTI1,
-    gcvFEATURE_HALTI2,
-    gcvFEATURE_SUPPORT_GCREGTX,
-    gcvFEATURE_2D_MIRROR_EXTENSION,
-    gcvFEATURE_TEXTURE_ASTC,
-    gcvFEATURE_TEXTURE_ASTC_DECODE_FIX,
-    gcvFEATURE_TEXTURE_ASTC_BASE_LOD_FIX,
-    gcvFEATURE_2D_SUPER_TILE_V1,
-    gcvFEATURE_2D_SUPER_TILE_V2,
-    gcvFEATURE_2D_SUPER_TILE_V3,
-    gcvFEATURE_2D_MULTI_SOURCE_BLT_EX2,
-    gcvFEATURE_NEW_RA,
-    gcvFEATURE_BUG_FIXED_IMPLICIT_PRIMITIVE_RESTART,
-    gcvFEATURE_PE_MULTI_RT_BLEND_ENABLE_CONTROL,
-    gcvFEATURE_SMALL_MSAA, /* An upgraded version of Fast MSAA */
-    gcvFEATURE_VERTEX_INST_ID_AS_ATTRIBUTE,
-    gcvFEATURE_DUAL_16,
-    gcvFEATURE_BRANCH_ON_IMMEDIATE_REG,
-    gcvFEATURE_2D_COMPRESSION,
-    gcvFEATURE_TPC_COMPRESSION,
-    gcvFEATURE_TPCV11_COMPRESSION,
-    gcvFEATURE_DEC_COMPRESSION,
-    gcvFEATURE_DEC300_COMPRESSION,
-    gcvFEATURE_DEC400_COMPRESSION,
-    /*Its a new core for DEC400 compression*/
-    gcvFEATURE_DEC400EX_COMPRESSION,
-    gcvFEATURE_DEC_TPC_COMPRESSION,
-    gcvFEATURE_DEC_COMPRESSION_TILE_NV12_8BIT,
-    gcvFEATURE_DEC_COMPRESSION_TILE_NV12_10BIT,
-    gcvFEATURE_2D_OPF_YUV_OUTPUT,
-    gcvFEATURE_2D_FILTERBLIT_A8_ALPHA,
-    gcvFEATURE_2D_MULTI_SRC_BLT_TO_UNIFIED_DST_RECT,
-    gcvFEATURE_2D_MULTI_SRC_BLT_BILINEAR_FILTER,
-    gcvFEATURE_2D_MULTI_SRC_BLT_1_5_ENHANCEMENT,
-    gcvFEATURE_V2_COMPRESSION_Z16_FIX,
-    gcvFEATURE_VERTEX_INST_ID_AS_INTEGER,
-    gcvFEATURE_2D_YUV_MODE,
-    gcvFEATURE_2D_CACHE_128B256BPERLINE,
-    gcvFEATURE_2D_SEPARATE_CACHE,
-    gcvFEATURE_2D_MAJOR_SUPER_TILE,
-    gcvFEATURE_2D_V4COMPRESSION,
-    gcvFEATURE_2D_VMSAA,
-    gcvFEATURE_2D_10BIT_OUTPUT_LINEAR,
-    gcvFEATURE_2D_YUV420_OUTPUT_LINEAR,
-    gcvFEATURE_ACE,
-    gcvFEATURE_COLOR_COMPRESSION,
-    gcvFEATURE_32BPP_COMPONENT_TEXTURE_CHANNEL_SWIZZLE,
-    gcvFEATURE_64BPP_HW_CLEAR_SUPPORT,
-    gcvFEATURE_TX_LERP_PRECISION_FIX,
-    gcvFEATURE_COMPRESSION_V2,
-    gcvFEATURE_MMU,
-    gcvFEATURE_COMPRESSION_V3,
-    gcvFEATURE_TX_DECOMPRESSOR,
-    gcvFEATURE_MRT_TILE_STATUS_BUFFER,
-    gcvFEATURE_COMPRESSION_V1,
-    gcvFEATURE_V1_COMPRESSION_Z16_DECOMPRESS_FIX,
-    gcvFEATURE_RTT,
-    gcvFEATURE_GENERIC_ATTRIB,
-    gcvFEATURE_2D_ONE_PASS_FILTER,
-    gcvFEATURE_2D_ONE_PASS_FILTER_TAP,
-    gcvFEATURE_2D_POST_FLIP,
-    gcvFEATURE_2D_PIXEL_ALIGNMENT,
-    gcvFEATURE_CORRECT_AUTO_DISABLE_COUNT,
-    gcvFEATURE_CORRECT_AUTO_DISABLE_COUNT_WIDTH,
-    gcvFEATURE_8K_RT,
-    gcvFEATURE_HALTI3,
-    gcvFEATURE_EEZ,
-    gcvFEATURE_INTEGER_SIGNEXT_FIX,
-    gcvFEATURE_PSOUTPUT_MAPPING,
-    gcvFEATURE_8K_RT_FIX,
-    gcvFEATURE_TX_TILE_STATUS_MAPPING,
-    gcvFEATURE_SRGB_RT_SUPPORT,
-    gcvFEATURE_TEXTURE_16K,
-    gcvFEATURE_PA_FARZCLIPPING_FIX,
-    gcvFEATURE_PE_DITHER_COLORMASK_FIX,
-    gcvFEATURE_ZSCALE_FIX,
-    gcvFEATURE_MULTI_PIXELPIPES,
-    gcvFEATURE_PIPE_CL,
-    gcvFEATURE_BUG_FIXES18,
-    gcvFEATURE_UNIFIED_SAMPLERS,
-    gcvFEATURE_CL_PS_WALKER,
-    gcvFEATURE_NEW_HZ,
-    gcvFEATURE_TX_FRAC_PRECISION_6BIT,
-    gcvFEATURE_SH_INSTRUCTION_PREFETCH,
-    gcvFEATURE_PROBE,
-    gcvFEATURE_SINGLE_PIPE_HALTI1,
-    gcvFEATURE_BUG_FIXES8, /* This HW feature is wrong, we can't use this to check integer branch!!!*/
-    gcvFEATURE_2D_ALL_QUAD,
-    gcvFEATURE_SEPARATE_SRC_DST,
-    gcvFEATURE_TX_HOR_ALIGN_SEL,
-    gcvFEATURE_HALTI4,
-    gcvFEATURE_MRT_FC_FIX,
-    gcvFEATURE_TESSELLATION,
-    gcvFEATURE_DRAW_INDIRECT,
-    gcvFEATURE_COMPUTE_INDIRECT,
-    gcvFEATURE_MSAA_TEXTURE,
-    gcvFEATURE_STENCIL_TEXTURE,
-    gcvFEATURE_S8_ONLY_RENDERING,
-    gcvFEATURE_D24S8_SAMPLE_STENCIL,
-    gcvFEATURE_ADVANCED_BLEND_MODE_PART0,
-    gcvFEATURE_RA_DEPTH_WRITE,
-    gcvFEATURE_RS_DS_DOWNSAMPLE_NATIVE_SUPPORT,
-    gcvFEATURE_S8_MSAA_COMPRESSION,
-    gcvFEATURE_MSAA_FRAGMENT_OPERATION,
-    gcvFEATURE_FE_START_VERTEX_SUPPORT,
-    gcvFEATURE_DIVISOR_STREAM_ADDR_FIX,
-    gcvFEATURE_ZERO_ATTRIB_SUPPORT,
-    gcvFEATURE_DANGLING_VERTEX_FIX,
-    gcvFEATURE_PE_DISABLE_COLOR_PIPE,
-    gcvFEATURE_FE_12bit_stride,
-    gcvFEATURE_TX_LOD_GUARDBAND,
-    gcvFEATURE_HAS_PRODUCTID,
-    gcvFEATURE_INTEGER32_FIX,
-    gcvFEATURE_TEXTURE_GATHER,
-    gcvFEATURE_IMG_INSTRUCTION,
-    gcvFEATURE_HELPER_INVOCATION,
-    gcvFEATURE_NO_USER_CSC,
-    gcvFEATURE_ANDROID_ONLY,
-    gcvFEATURE_V2_MSAA_COHERENCY_FIX,
-    gcvFEATURE_BLOCK_SIZE_16x16,
-    gcvFEATURE_TX_SUPPORT_DEC,
-    gcvFEATURE_RSBLT_MSAA_DECOMPRESSION,
-    gcvFEATURE_TILEFILLER_32TILE_ALIGNED,
-    gcvFEATURE_GEOMETRY_SHADER,
-    gcvFEATURE_HALTI5,
-    gcvFEATURE_PIPELINE_32_ATTRIBUTES,
-    gcvFEATURE_USC,
-    gcvFEATURE_CUBEMAP_ARRAY,
-    gcvFEATURE_TX_DESCRIPTOR,
-    gcvFEATURE_SEPARATE_RT_CTRL,
-    gcvFEATURE_RENDER_ARRAY,
-    gcvFEATURE_BLT_ENGINE,
-    gcvFEATURE_TEXTURE_BUFFER,
-    gcvFEATURE_GS_SUPPORT_EMIT,
-    gcvFEATURE_SAMPLER_BASE_OFFSET,
-    gcvFEATURE_IMAGE_OUT_BOUNDARY_FIX,
-    gcvFEATURE_TX_BORDER_CLAMP,
-    gcvFEATURE_MSAA_SHADING,
-    gcvFEATURE_ADVANCED_SH_INST,
-    gcvFEATURE_LOD_FIX_FOR_BASELEVEL,
-    gcvFEATURE_MULTIDRAW_INDIRECT,
-    gcvFEATURE_DRAW_ELEMENTS_BASE_VERTEX,
-    gcvFEATURE_NEW_STEERING_AND_ICACHE_FLUSH, /* Steering base on register base. Trigger-style Icache flush state. */
-    gcvFEATURE_PE_DITHER_FIX2,
-    gcvFEATURE_INDEX_FETCH_FIX,
-    gcvFEATURE_TEX_BASELOD,
-    gcvFEATURE_TEX_SEAMLESS_CUBE,
-    gcvFEATURE_TEX_ETC2,
-    gcvFEATURE_TEX_CUBE_BORDER_LOD,
-    gcvFEATURE_FE_ALLOW_STALL_PREFETCH_ENG,
-    gcvFEATURE_TX_8BPP_TS_FIX,
-    gcvFEATURE_HW_TFB,
-    gcvFEATURE_COMPRESSION_V4,
-    gcvFEATURE_FENCE_32BIT,
-    gcvFEATURE_FENCE_64BIT,
-    gcvFEATURE_R8_UNORM,
-    gcvFEATURE_TX_DEFAULT_VALUE_FIX,
-    gcvFEATURE_TX_8bit_UVFrac,
-    gcvFEATURE_TX_MIPFILTER_NONE_FIX,
-    gcvFEATURE_MC_STENCIL_CTRL,
-    gcvFEATURE_DEPTH_MATH_FIX,
-    gcvFEATURE_PE_B2B_PIXEL_FIX,
-    gcvFEATURE_TEXTURE_GATHER_OFFSETS,
-    gcvFEATURE_TEX_CACHE_FLUSH_FIX,
-    gcvFEATURE_WIDELINE_HELPER_FIX,
-    gcvFEATURE_LINE_DIAMOND_RULE_FIX,
-    gcvFEATURE_MULTIGPU_SYNC_V2,
-    gcvFEATURE_DRAW_ID,
-    gcvFEATURE_SNAPPAGE_CMD,
-    gcvFEATURE_COMMAND_PREFETCH,
-    gcvFEATURE_SAMPLEPOS_SWIZZLE_FIX,
-    gcvFEATURE_SELECTMAP_SRC0_SWIZZLE_FIX,
-    gcvFEATURE_LOADATTR_OOB_FIX,
-    gcvFEATURE_RA_DEPTH_WRITE_MSAA1X_FIX,
-    gcvFEATURE_MRT_8BIT_DUAL_PIPE_FIX,
-    gcvFEATURE_BUG_FIXES1,
-    gcvFEATURE_MULTI_SOURCE_BLT,
-    gcvFEATURE_ZCOMPRESSION,
-    gcvFEATURE_DITHER_AND_FILTER_PLUS_ALPHA_2D,
-    gcvFEATURE_ONE_PASS_2D_FILTER,
-    gcvFEATURE_TX_FILTER,
-    gcvFEATURE_CHIPENABLE_LINK,
-    gcvFEATURE_TEXTURE_BIAS_LOD_FIX,
-    gcvFEATURE_USE_GL_Z,
-    gcvFEATURE_SUPPORT_INTEGER,
-    /* PARTLY_SUPPORT_INTEGER_BRANCH:
-    **      chips can support all integer types for compare instructions, e.g, CMP, SELECT.
-    ** FULLLY_SUPPORT_INTEGER_BRANCH:
-    **      chips can support all integer types for JMP instruction.
-    ** If PARTLY_SUPPORT_INTEGER_BRANCH is TRUE but FULLLY_SUPPORT_INTEGER_BRANCH is FALSE,
-    ** then this chip can only support INT32/UINT32 JMP instruction.
-    */
-    gcvFEATURE_PARTLY_SUPPORT_INTEGER_BRANCH,
-    gcvFEATURE_FULLLY_SUPPORT_INTEGER_BRANCH,
-    gcvFEATURE_SUPPORT_INTEGER_ATTRIBUTE,
-    gcvFEATURE_SUPPORT_MOVAI,
-    gcvFEATURE_NEED_FIX_FOR_CL_X,
-    gcvFEATURE_NEED_FIX_FOR_CL_XE,
-    gcvFEATURE_HAS_OUTPUT_COUNT_FIX,
-    gcvFEATURE_VARYING_PACKING_LIMITATION,
-    gcvFEATURE_HIGHP_VARYING_SHIFT,
-    gcvFEATURE_BUG_FIXES2,
-    gcvFEATURE_64K_L2_CACHE,
-    gcvFEATURE_128BTILE,
-    gcvFEATURE_ADVANCED_BLEND_OPT,
-    gcvFEATURE_SNAPPAGE_CMD_FIX,
-    gcvFEATURE_L2_CACHE_FOR_2D_420,
-    gcvFEATURE_TILE_STATUS_2BITS,
-    gcvFEATURE_EXTRA_SHADER_INSTRUCTIONS0,
-    gcvFEATURE_EXTRA_SHADER_INSTRUCTIONS1,
-    gcvFEATURE_EXTRA_SHADER_INSTRUCTIONS2,
-    gcvFEATURE_MEDIUM_PRECISION,
-    gcvFEATURE_FE20_BIT_INDEX,
-    gcvFEATURE_BUG_FIXES4,
-    gcvFEATURE_BUG_FIXES12,
-    gcvFEATURE_VMSAA,
-    gcvFEATURE_ROBUST_ATOMIC,
-    gcvFEATURE_32F_COLORMASK_FIX,
-    gcvFEATURE_NEW_GPIPE,
-    gcvFEATURE_RS_NEW_BASEADDR,
-    gcvFEATURE_TX_DXT,
-    gcvFEATURE_SH_FLAT_INTERPOLATION_DUAL16_FIX,
-    gcvFEATURE_EVIS,
-    gcvFEATURE_SH_SUPPORT_V4,
-    gcvFEATURE_SH_SUPPORT_ALPHA_KILL,
-    gcvFEATURE_PE_NO_ALPHA_TEST,
-    gcvFEATURE_SH_SNAP2PAGE_MAXPAGES_FIX,
-    gcvFEATURE_USC_FULLCACHE_FIX,
-    gcvFEATURE_PE_64bit_FENCE_FIX,
-    gcvFEATURE_BLT_8bit_256TILE_FC_FIX,
-    gcvFEATURE_PE_RGBA16I_FIX,
-    gcvFEATURE_BLT_64bpp_MASKED_CLEAR_FIX,
-    gcvFEATURE_SH_PSO_MSAA1x_FIX,
-    gcvFEATURE_USC_ATOMIC_FIX,
-    gcvFEATURE_INDEX_CONST_ON_B0,
-    gcvFEATURE_SH_NO_ONECONST_LIMIT,
-    gcvFEATURE_EVIS_NO_ABSDIFF,
-    gcvFEATURE_EVIS_NO_BITREPLACE,
-    gcvFEATURE_EVIS_NO_BOXFILTER,
-    gcvFEATURE_EVIS_NO_CORDIAC,
-    gcvFEATURE_EVIS_NO_DP32,
-    gcvFEATURE_EVIS_NO_FILTER,
-    gcvFEATURE_EVIS_NO_IADD,
-    gcvFEATURE_EVIS_NO_SELECTADD,
-    gcvFEATURE_EVIS_LERP_7OUTPUT,
-    gcvFEATURE_EVIS_ACCSQ_8OUTPUT,
-    gcvFEATURE_ROBUSTNESS,
-    gcvFEATURE_SECURITY,
-    gcvFEATURE_TX_YUV_ASSEMBLER_10BIT,
-    gcvFEATURE_USC_GOS_ADDR_FIX,
-    gcvFEATURE_SUPPORT_MSAA2X,
-    gcvFEATURE_TX_DESC_CACHE_CLOCKGATE_FIX,
-    gcvFEATURE_TX_INTEGER_COORDINATE,
-    gcvFEATURE_PSIO_SAMPLEMASK_IN_R0ZW_FIX,
-    gcvFEATURE_MULTI_CORE_BLOCK_SET_CONFIG,
-    gcvFEATURE_SH_IMG_LDST_ON_TEMP,
-    gcvFEATURE_TX_INTEGER_COORDINATE_V2,
-    gcvFEATURE_COMPUTE_ONLY,
-    gcvFEATURE_SH_IMG_LDST_CLAMP,
-    gcvFEATURE_SH_ICACHE_ALLOC_COUNT_FIX,
-    gcvFEATURE_MSAA_OQ_FIX,
-    gcvFEATURE_PE_ENHANCEMENTS2,
-    gcvFEATURE_PSIO_MSAA_CL_FIX,
-    gcvFEATURE_FE_NEED_DUMMYDRAW,
-    gcvFEATURE_MULTI_CLUSTER,
-    gcvFEATURE_PSIO_INTERLOCK,
-    gcvFEATURE_BLIT_COMPRESS_DEST,
-    gcvFEATURE_SH_MULTI_WG_PACK,
-    gcvFEATURE_FE_ROBUST_FIX,
-    gcvFEATURE_TX_ASTC_MULTISLICE_FIX,
-    gcvFEATURE_PSIO_DUAL16_32bpc_FIX,
-    gcvFEATURE_LS_SUPPORT_PER_COMP_DEPENDENCY,
-    gcvFEATURE_COMPRESSION_DEC400,
-    gcvFEATURE_SH_TEXLD_U_FIX,
-    gcvFEATURE_TX_FLUSH_L1CACHE,
-    gcvFEATURE_USC_DEFER_FILL_FIX,
-    gcvFEATURE_MC_FCCACHE_BYTEMASK,
-    gcvFEATURE_SH_MULTI_WG_PACK_FIX,
-    gcvFEATURE_FE_PATCHLIST_FETCH_FIX,
-    gcvFEATURE_RA_CG_FIX,
-    gcvFEATURE_EVIS_VX2,
-    gcvFEATURE_SH_HALF_DEPENDENCY_FIX,
-    gcvFEATURE_FE_BASEINSTANCE,
-    gcvFEATURE_FE_COMPUREINDIRECT_SKIP_UNIFORM,
-    gcvFEATURE_SH_CLOCK_GATE_FIX,
-    gcvFEATURE_GPIPE_CLOCK_GATE_FIX,
-    gcvFEATURE_TP_ENGINE,
-    gcvFEATURE_TX_BORDER_CLAMP_FIX,
-    gcvFEATURE_SH_IMAGE_LD_LAST_PIXEL_FIX,
-    gcvFEATURE_MULTI_CORE_BLOCK_SET_CONFIG2,
-    gcvFEATURE_MULTIGPU_SYNC_V3,
-    gcvFEATURE_PE_VMSAA_COVERAGE_CACHE_FIX,
-    gcvFEATURE_SECURITY_AHB,
-    gcvFEATURE_TX_LERP_LESS_BIT,
-    gcvFEATURE_SMALL_BATCH,
-    gcvFEATURE_SH_IDIV0_SWZL_EHS,
-    gcvFEATURE_SH_CMPLX,
-    gcvFEATURE_VIP_V7,
-    gcvFEATURE_SH_GM_ENDIAN,
-    gcvFEATURE_SH_GM_USC_UNALLOC,
-    gcvFEATURE_SH_END_OF_BB,
-    gcvFEATURE_ASYNC_BLIT,
-    gcvFEATURE_ASYNC_FE_FENCE_FIX,
-    gcvFEATURE_PSCS_THROTTLE,
-    gcvFEATURE_SEPARATE_LS,
-    gcvFEATURE_PA_VARYING_COMPONENT_TOGGLE_FIX,
-    gcvFEATURE_TX_MULTISAMPLER_FC_FIX,
-    gcvFEATURE_WIDELINE_TRIANGLE_EMU,
-    gcvFEATURE_FENCE,
-    gcvFEATURE_MCFE,
-    gcvFEATURE_NN_INTERLEAVE8,
-    gcvFEATURE_TP_REORDER,
-    gcvFEATURE_TP_RTNE,
-    gcvFEATURE_TP_LRN,
-    gcvFEATURE_TP_ROI_POOLING,
-    gcvFEATURE_TP_MAX_POOLING_STRIDE1,
-    gcvFEATURE_NN_BRICK_MODE,
-    gcvFEATURE_NN_BORDER_MODE,
-    gcvFEATURE_NN_FP16_ALU,
-    gcvFEATURE_NN_BF16_ALU,
-    gcvFEATURE_NN_INT16_ALU,
-    gcvFEATURE_NN_ZDP3,
-    gcvFEATURE_NN_ZDP6,
-    gcvFEATURE_PE_DEPTH_ONLY_OQFIX,
-    gcvFEATURE_TX_SNORM_SUPPORT,
-    gcvFEATURE_HWMANAGED_LS,
-    gcvFEATURE_SH_SCATTER_GATHER,
-    gcvFEATURE_NN_POWER_ISOLATION,
-    gcvFEATURE_SWTILING_PHASE1,
-    gcvFEATURE_SWTILING_PHASE2,
-    gcvFEATURE_SWTILING_PHASE3,
-    gcvFEATURE_TF_QUANTIZATION,
-    gcvFEATURE_NN_XYDP9,
-    gcvFEATURE_TP_SIMPLE_INT16,
-    gcvFEATURE_TP_REAL_INT16,
-    gcvFEATURE_NN_FIRST_PIXEL_POOLING,
-    gcvFEATURE_NN_STRIDE_SUPPORT,
-    gcvFEATURE_NN_XYDP6,
-    gcvFEATURE_NN_XYDP0,
-    gcvFEATURE_TP_REORDER_FIX,
-    gcvFEATURE_NN_CONV1x1_PERF_FIX,
-    gcvFEATURE_NN_CACHELINE_MODE_PERF_FIX,
-    gcvFEATURE_NN_PER3DTILE_BUBBLE_FIX,
-    gcvFEATURE_SH_IO_CG_FIX,
-    gcvFEATURE_USC_STAY_LRU,
-    gcvFEATURE_NN_NONZERO_MIRROR_BORDER,
-    gcvFEATURE_NN_COEF_DECOMPRESS_PERF2X,
-    gcvFEATURE_4BIT_INPUT,
-    gcvFEATURE_COEF_COMPRESSION_ENHANCEMENT,
-    gcvFEATURE_NN_ZDP3_NO_COMPRESS_FIX,
-    gcvFEATURE_NN_ASYNC_COPY_PERF_FIX,
-    gcvFEATURE_OCB_COUNTER,
-    gcvFEATURE_NN_ZXDP3_KERNEL_READ_CONFLICT_FIX,
-    gcvFEATURE_NN_FULLCACHE_KERNEL_INTERLEAVE_FIX,
-    gcvFEATURE_DR_JD_DIFF_CONDITION_FOR_CACHELINE_MODE_PRE_FIX,
-    gcvFEATURE_USC_BOTTLENECK_FIX,
-    gcvFEATURE_OCB_REMAP_PHYSICAL_ADDRESS,
-    gcvFEATURE_NN_SLICE_PADDING_TO_64BYTE_ALIGN,
-    gcvFEATURE_NN_DW_1x1_CONV_MERGE,
-    gcvFEATURE_TP_REORDER_LAYER_SUSPEND_FIX,
-    gcvFEATURE_KERNEL_VIP_SRAM_READ_BW_LIMITATION_FIX,
-    gcvFEATURE_IMG_POP_PIPELINE_PAUSE_FIX,
-    gcvFEATURE_NN_SLOW_OUTPUT,
-    gcvFEATURE_NO_NARROW_POST_PROCESS_PIPE,
-    gcvFEATURE_TP_NN_PROBE,
-    gcvFEATURE_TP_23BITS_POST_MULTIPLIER,
-    gcvFEATURE_NN_TRANSPOSE,
-    gcvFEATURE_OUTIMAGE_X_BITWIDTH_LIMIT_FOR_NN_TRANSPOSE_FIX,
-    gcvFEATURE_TP_BFLOAT16,
-    gcvFEATURE_EVIS2_FLOP_RESET_FIX,
-    gcvFEATURE_USC_ASYNC_CP_RTN_FLOP_RESET_FIX,
-    gcvFEATURE_USC_EVICT_CTRL_FIFO_FLOP_RESET_FIX,
-
-    gcvFEATURE_IMAGE_LS_NO_FULLMASK_FIX,
-    gcvFEATURE_BLT_YUV_OUTPUT,
-    gcvFEATURE_PE_TILE_CACHE_FLUSH_FIX,
-    gcvFEATURE_SH_ROBUSTNESS_FIX,
-    gcvFEATURE_USC_ATOMIC_FIX2,
-    gcvFEATURE_MULTIVIEW_RENDER,
-    gcvFEATURE_FE_DRAW_DIRECT,
-    gcvFEATURE_TX_VKBORDER_MODE,
-    gcvFEATURE_TX_UNNORMALIZED_COORD,
-    gcvFEATURE_VG_IMAGE_16K,
-    gcvFEATURE_MULTICORE_CONFIG,
-    gcvFEATURE_PA_LINECLIP_FIX,
-    gcvFEATURE_NN_ENGINE,
-    gcvFEATURE_NN_ASYNC_COPY_MERGE_FIX,
-    gcvFEATURE_NN_CONVOUT_FIFO_DEPTH_FIX,
-    gcvFEATURE_NN_SMALLBATCH_PHASE1,
-    gcvFEATURE_TP_SMALLBATCH_PHASE1,
-    gcvFEATURE_VIP_SCALER,
-    gcvFEATURE_TX_8bit_UVFrac_ROUNDING_FIX,
-    gcvFEATURE_NN_REQ_SLOWARBITRATION_FIX,
-    gcvFEATUER_IMAGE_PARTIAL_CACHE,
-    gcvFEATURE_FULLCACHE_KERNELHEAD_FIX,
-    gcvFEATURE_NN_SINGLEPORT_ACCUMBUFFER,
-    gcvFEATURE_NN_SMALLBATCH,
-    gcvFEATURE_TP_SMALLBATCH,
-    gcvFEATURE_NN_ZDP_INIMAGE_SIZE_FIX,
-    gcvFEATURE_HI_REORDER_FIX,
-    gcvFEATURE_TP_COEF_COMPRESSION_ENHANCEMENT,
-    gcvFEATURE_NN_DEPTHWISE_SUPPORT,
-    gcvFEATURE_IMAGE_NOT_PACKED_IN_SRAM_FIX,
-    gcvFEATURE_IDLE_BEFORE_FLUSH_COMPLETE_FIX,
-    gcvFEATURE_NO_FLUSH_USC_FIX,
-    gcvFEATURE_COEF_DELTA_CORD_OVERFLOW_ZRL_8BIT_FIX,
-    gcvFEATURE_XY_OFFSET_LIMITATION_FIX,
-    gcvFEATURE_USC_INVALIDATE_CACHE_LINE_FIX,
-    gcvFEATURE_LOW_EFFICIENCY_OF_ID_WRITE_IMGBUF_FIX,
-    gcvFEATURE_KERNEL_PER_CORE_LESS_THAN_THIRD_COEF_BUFF_DEPTH_FIX,
-    gcvFEATURE_NN_PER_CHANNEL_QUANT,
-    gcvFEATURE_NN_NO_Z_LOCATION_OFFSET,
-    gcvFEATURE_NN_PRELU,
-    gcvFEATURE_NN_KERNEL_SIZE_WASTE_IN_PARTIAL_MODE_FIX,
-    gcvFEATURE_INCORRECT_WR_REQ_TO_USC_BETWEEN_REORDER_AND_NORMAL_LAYER_FIX,
-    gcvFEATURE_VIP_DEC400,
-    gcvFEATURE_MAX_POINTSIZE_CLAMP,
-    gcvFEATURE_2D_FAST_CLEAR, /* For tilestatus Fast Clear feature*/
-    gcvFEATURE_NEGATIVE_POST_SHIFT_FIX,
-    gcvFEATURE_NN_COMMAND_KERNEL_REQUEST_CONFICT_FIX,
-    gcvFEATURE_DDR_BURST_LEN_256B,
-    gcvFEATURE_NN_LEAKY_RELU,
-    gcvFEATURE_NN_NATIVE_STRIDE_TWO,
-    gcvFEATURE_NN_TENSOR_ADD,
-    gcvFEATURE_VIP_SCALER_4K,
-    gcvFEATURE_NN_PER_CHANNEL_QUANT_ASYM,
-    gcvFEATURE_PE_A8B8G8R8, /* For PE support A8B8G8R8 format feature*/
-
-    /* Insert features above this comment only. */
-    gcvFEATURE_COUNT                /* Not a feature. */
-}
-gceFEATURE;
 
 /* dummy draw type.*/
 typedef enum _gceDUMMY_DRAW_TYPE
@@ -728,128 +142,6 @@ typedef enum _gceFRAMEINFO_OP
 }
 gceFRAMEINFO_OP;
 
-
-/* Chip Power Status. */
-typedef enum _gceCHIPPOWERSTATE
-{
-    gcvPOWER_INVALID = -1,
-
-    /* Global/base states. */
-    gcvPOWER_ON = 0,
-    gcvPOWER_IDLE,
-    gcvPOWER_SUSPEND,
-    gcvPOWER_OFF,
-
-    /* Power on but not global or broadcast. */
-    gcvPOWER_ON_AUTO,
-
-    /* Broadcast states. */
-    gcvPOWER_FLAG_BROADCAST    = 0x10,
-    gcvPOWER_IDLE_BROADCAST    = gcvPOWER_IDLE    | gcvPOWER_FLAG_BROADCAST,
-    gcvPOWER_SUSPEND_BROADCAST = gcvPOWER_SUSPEND | gcvPOWER_FLAG_BROADCAST,
-    gcvPOWER_OFF_BROADCAST     = gcvPOWER_OFF     | gcvPOWER_FLAG_BROADCAST,
-
-
-    /* Timeout states. */
-    gcvPOWER_FLAG_TIMEOUT      = 0x20,
-    gcvPOWER_IDLE_TIMEOUT      = gcvPOWER_IDLE    | gcvPOWER_FLAG_TIMEOUT,
-    gcvPOWER_SUSPEND_TIMEOUT   = gcvPOWER_SUSPEND | gcvPOWER_FLAG_TIMEOUT,
-    gcvPOWER_OFF_TIMEOUT       = gcvPOWER_OFF     | gcvPOWER_FLAG_TIMEOUT,
-
-}
-gceCHIPPOWERSTATE;
-
-/* CPU cache operations */
-typedef enum _gceCACHEOPERATION
-{
-    gcvCACHE_CLEAN      = 0x01, /* Flush CPU cache to mem */
-    gcvCACHE_INVALIDATE = 0x02, /* Invalidte CPU cache */
-    gcvCACHE_FLUSH      = gcvCACHE_CLEAN  | gcvCACHE_INVALIDATE, /* Both flush & invalidate */
-    gcvCACHE_MEMORY_BARRIER = 0x04
-}
-gceCACHEOPERATION;
-
-/* Surface types. */
-typedef enum _gceSURF_TYPE
-{
-    gcvSURF_TYPE_UNKNOWN = 0,
-    gcvSURF_INDEX,
-    gcvSURF_VERTEX,
-    gcvSURF_TEXTURE,
-    gcvSURF_RENDER_TARGET,
-    gcvSURF_DEPTH,
-    gcvSURF_BITMAP,
-    gcvSURF_TILE_STATUS,
-    gcvSURF_IMAGE,
-    gcvSURF_MASK,
-    gcvSURF_SCISSOR,
-    gcvSURF_HIERARCHICAL_DEPTH,
-    gcvSURF_ICACHE,
-    gcvSURF_TXDESC,
-    gcvSURF_FENCE,
-    gcvSURF_TFBHEADER,
-    gcvSURF_NUM_TYPES, /* Make sure this is the last one! */
-
-    /* Combinations. */
-    gcvSURF_CMA_LIMIT               = 0x80000000,
-    gcvSURF_NO_TILE_STATUS          = 0x100,
-    gcvSURF_NO_VIDMEM               = 0x200, /* Used to allocate surfaces with no underlying vidmem node.
-                                                   In Android, vidmem node is allocated by another process. */
-    gcvSURF_CACHEABLE               = 0x400, /* Used to allocate a cacheable surface */
-    gcvSURF_TILE_RLV_FENCE          = 0x800, /* create texture fence as tile */
-    gcvSURF_TILE_STATUS_DIRTY       = 0x1000, /* Init tile status to all dirty */
-    gcvSURF_LINEAR                  = 0x2000,
-    gcvSURF_CREATE_AS_TEXTURE       = 0x4000, /* create it as a texture */
-    gcvSURF_PROTECTED_CONTENT       = 0x8000, /* create it as content protected */
-    gcvSURF_CREATE_AS_DISPLAYBUFFER = 0x10000, /*create it as a display buffer surface */
-    gcvSURF_CONTIGUOUS              = 0x20000, /*create it as contiguous */
-    gcvSURF_NO_COMPRESSION          = 0x40000, /* Create it as no compression, valid on when it has tile status. */
-    gcvSURF_DEC                     = 0x80000, /* Surface is DEC compressed */
-    gcvSURF_NO_HZ                   = 0x100000,
-    gcvSURF_3D                      = 0x200000, /* It's 3d surface */
-    gcvSURF_DMABUF_EXPORTABLE       = 0x400000, /* master node can be exported as dma-buf fd */
-    gcvSURF_CACHE_MODE_128          = 0x800000,
-
-    gcvSURF_TEXTURE_LINEAR               = gcvSURF_TEXTURE
-                                         | gcvSURF_LINEAR,
-
-    gcvSURF_RENDER_TARGET_LINEAR         = gcvSURF_RENDER_TARGET
-                                         | gcvSURF_LINEAR,
-
-    gcvSURF_RENDER_TARGET_NO_TILE_STATUS = gcvSURF_RENDER_TARGET
-                                         | gcvSURF_NO_TILE_STATUS,
-
-    gcvSURF_RENDER_TARGET_NO_COMPRESSION = gcvSURF_RENDER_TARGET
-                                         | gcvSURF_NO_COMPRESSION,
-
-    gcvSURF_RENDER_TARGET_TS_DIRTY       = gcvSURF_RENDER_TARGET
-                                         | gcvSURF_TILE_STATUS_DIRTY,
-
-    gcvSURF_DEPTH_NO_TILE_STATUS         = gcvSURF_DEPTH
-                                         | gcvSURF_NO_TILE_STATUS,
-
-    gcvSURF_DEPTH_TS_DIRTY               = gcvSURF_DEPTH
-                                         | gcvSURF_TILE_STATUS_DIRTY,
-
-    /* Supported surface types with no vidmem node. */
-    gcvSURF_BITMAP_NO_VIDMEM             = gcvSURF_BITMAP
-                                         | gcvSURF_NO_VIDMEM,
-
-    gcvSURF_TEXTURE_NO_VIDMEM            = gcvSURF_TEXTURE
-                                         | gcvSURF_NO_VIDMEM,
-
-    /* Cacheable surface types with no vidmem node. */
-    gcvSURF_CACHEABLE_BITMAP_NO_VIDMEM   = gcvSURF_BITMAP_NO_VIDMEM
-                                         | gcvSURF_CACHEABLE,
-
-    gcvSURF_CACHEABLE_BITMAP             = gcvSURF_BITMAP
-                                         | gcvSURF_CACHEABLE,
-
-    gcvSURF_TEXTURE_3D                  = gcvSURF_TEXTURE
-                                         | gcvSURF_3D
-}
-gceSURF_TYPE;
-
 typedef enum _gceSURF_USAGE
 {
     gcvSURF_USAGE_UNKNOWN,
@@ -912,427 +204,6 @@ typedef enum _gceMIPMAP_IMAGE_FORMAT
     gcvUNKNOWN_MIPMAP_IMAGE_FORMAT  = -2
 }
 gceMIPMAP_IMAGE_FORMAT;
-
-/* Surface formats.
-** Name rules is from MSB->LSB.
-*/
-typedef enum _gceSURF_FORMAT
-{
-    /* Unknown format. */
-    gcvSURF_UNKNOWN             = 0,
-
-    /* Palettized formats. */
-    gcvSURF_INDEX1              = 100,
-    gcvSURF_INDEX4,
-    gcvSURF_INDEX8,
-#if gcdVG_ONLY
-    gcvSURF_INDEX2,
-#endif
-
-    /* RGB formats. */
-    gcvSURF_A2R2G2B2            = 200,
-    gcvSURF_R3G3B2,
-    gcvSURF_A8R3G3B2,
-    gcvSURF_X4R4G4B4,
-    gcvSURF_A4R4G4B4,
-    gcvSURF_R4G4B4A4,
-    gcvSURF_X1R5G5B5,
-    gcvSURF_A1R5G5B5,
-    gcvSURF_R5G5B5A1,
-    gcvSURF_R5G6B5,
-    gcvSURF_R8G8B8,
-    gcvSURF_X8R8G8B8,
-    gcvSURF_A8R8G8B8,
-    gcvSURF_R8G8B8A8,
-    gcvSURF_G8R8G8B8,
-    gcvSURF_R8G8B8G8,
-    gcvSURF_X2R10G10B10,
-    gcvSURF_A2R10G10B10,
-    gcvSURF_R10G10B10A2,
-    gcvSURF_X12R12G12B12,
-    gcvSURF_A12R12G12B12,
-    gcvSURF_X16R16G16B16,
-    gcvSURF_A16R16G16B16,
-    gcvSURF_A32R32G32B32,
-    gcvSURF_R8G8B8X8,
-    gcvSURF_R5G5B5X1,
-    gcvSURF_R4G4B4X4,
-    gcvSURF_X16R16G16B16_2_A8R8G8B8,
-    gcvSURF_A16R16G16B16_2_A8R8G8B8,
-    gcvSURF_A32R32G32B32_2_G32R32F,
-    gcvSURF_A32R32G32B32_4_A8R8G8B8,
-    /* BGR formats. */
-    gcvSURF_A4B4G4R4            = 300,
-    gcvSURF_A1B5G5R5,
-    gcvSURF_B5G6R5,
-    gcvSURF_B8G8R8,
-    gcvSURF_B16G16R16,
-    gcvSURF_X8B8G8R8,
-    gcvSURF_A8B8G8R8,
-    gcvSURF_A2B10G10R10,
-    gcvSURF_X16B16G16R16,
-    gcvSURF_A16B16G16R16,
-    gcvSURF_B32G32R32,
-    gcvSURF_X32B32G32R32,
-    gcvSURF_A32B32G32R32,
-    gcvSURF_B4G4R4A4,
-    gcvSURF_B5G5R5A1,
-    gcvSURF_B8G8R8X8,
-    gcvSURF_B8G8R8A8,
-    gcvSURF_B10G10R10A2,
-    gcvSURF_X4B4G4R4,
-    gcvSURF_X1B5G5R5,
-    gcvSURF_B4G4R4X4,
-    gcvSURF_B5G5R5X1,
-    gcvSURF_X2B10G10R10,
-    gcvSURF_B8G8R8_SNORM,
-    gcvSURF_X8B8G8R8_SNORM,
-    gcvSURF_A8B8G8R8_SNORM,
-    gcvSURF_A8B12G12R12_2_A8R8G8B8,
-
-    /* Compressed formats. */
-    gcvSURF_DXT1                = 400,
-    gcvSURF_DXT2,
-    gcvSURF_DXT3,
-    gcvSURF_DXT4,
-    gcvSURF_DXT5,
-    gcvSURF_CXV8U8,
-    gcvSURF_ETC1,
-    gcvSURF_R11_EAC,
-    gcvSURF_SIGNED_R11_EAC,
-    gcvSURF_RG11_EAC,
-    gcvSURF_SIGNED_RG11_EAC,
-    gcvSURF_RGB8_ETC2,
-    gcvSURF_SRGB8_ETC2,
-    gcvSURF_RGB8_PUNCHTHROUGH_ALPHA1_ETC2,
-    gcvSURF_SRGB8_PUNCHTHROUGH_ALPHA1_ETC2,
-    gcvSURF_RGBA8_ETC2_EAC,
-    gcvSURF_SRGB8_ALPHA8_ETC2_EAC,
-
-    /* YUV formats. */
-    gcvSURF_YUY2                = 500,
-    gcvSURF_UYVY,
-    gcvSURF_YV12,
-    gcvSURF_I420,
-    gcvSURF_NV12,
-    gcvSURF_NV21,
-    gcvSURF_NV16,
-    gcvSURF_NV61,
-    gcvSURF_YVYU,
-    gcvSURF_VYUY,
-    gcvSURF_AYUV,
-    gcvSURF_YUV420_10_ST,
-    gcvSURF_YUV420_TILE_ST,
-    gcvSURF_YUV420_TILE_10_ST,
-    gcvSURF_NV12_10BIT,
-    gcvSURF_NV21_10BIT,
-    gcvSURF_NV16_10BIT,
-    gcvSURF_NV61_10BIT,
-    gcvSURF_P010,
-    gcvSURF_P010_LSB,
-    gcvSURF_I010,
-#if gcdVG_ONLY
-    gcvSURF_AYUY2,
-    gcvSURF_ANV12,
-    gcvSURF_ANV16,
-    gcvSURF_AUYVY,
-    gcvSURF_YV16,
-#endif
-
-    /* Depth formats. */
-    gcvSURF_D16                 = 600,
-    gcvSURF_D24S8,
-    gcvSURF_D32,
-    gcvSURF_D24X8,
-    gcvSURF_D32F,
-    gcvSURF_S8D32F,
-    gcvSURF_S8D32F_1_G32R32F,
-    gcvSURF_S8D32F_2_A8R8G8B8,
-    gcvSURF_D24S8_1_A8R8G8B8,
-    gcvSURF_S8,
-    gcvSURF_X24S8,
-    gcvSURF_X24S8_1_A8R8G8B8,
-
-    /* Alpha formats. */
-    gcvSURF_A4                  = 700,
-    gcvSURF_A8,
-    gcvSURF_A12,
-    gcvSURF_A16,
-    gcvSURF_A32,
-    gcvSURF_A1,
-
-    /* Luminance formats. */
-    gcvSURF_L4                  = 800,
-    gcvSURF_L8,
-    gcvSURF_L12,
-    gcvSURF_L16,
-    gcvSURF_L32,
-    gcvSURF_L1,
-    gcvSURF_L8_RAW,
-
-    /* Alpha/Luminance formats. */
-    gcvSURF_A4L4                = 900,
-    gcvSURF_A2L6,
-    gcvSURF_A8L8,
-    gcvSURF_A4L12,
-    gcvSURF_A12L12,
-    gcvSURF_A16L16,
-
-    gcvSURF_A8L8_1_A8R8G8B8,
-
-    gcvSURF_A8L8_RAW,
-
-    /* Bump formats. */
-    gcvSURF_L6V5U5              = 1000,
-    gcvSURF_V8U8,
-    gcvSURF_X8L8V8U8,
-    gcvSURF_Q8W8V8U8,
-    gcvSURF_A2W10V10U10,
-    gcvSURF_V16U16,
-    gcvSURF_Q16W16V16U16,
-
-    /* R/RG/RA formats. */
-    gcvSURF_R8                  = 1100,
-    gcvSURF_X8R8,
-    gcvSURF_G8R8,
-    gcvSURF_X8G8R8,
-    gcvSURF_A8R8,
-    gcvSURF_R16,
-    gcvSURF_X16R16,
-    gcvSURF_G16R16,
-    gcvSURF_X16G16R16,
-    gcvSURF_A16R16,
-    gcvSURF_R32,
-    gcvSURF_X32R32,
-    gcvSURF_G32R32,
-    gcvSURF_X32G32R32,
-    gcvSURF_A32R32,
-    gcvSURF_RG16,
-    gcvSURF_R8_SNORM,
-    gcvSURF_G8R8_SNORM,
-
-    gcvSURF_R8_1_X8R8G8B8,
-    gcvSURF_G8R8_1_X8R8G8B8,
-
-    /* Floating point formats. */
-    gcvSURF_R16F                = 1200,
-    gcvSURF_X16R16F,
-    gcvSURF_G16R16F,
-    gcvSURF_X16G16R16F,
-    gcvSURF_B16G16R16F,
-    gcvSURF_X16B16G16R16F,
-    gcvSURF_A16B16G16R16F,
-    gcvSURF_R32F,
-    gcvSURF_X32R32F,
-    gcvSURF_G32R32F,
-    gcvSURF_X32G32R32F,
-    gcvSURF_B32G32R32F,
-    gcvSURF_X32B32G32R32F,
-    gcvSURF_A32B32G32R32F,
-    gcvSURF_A16F,
-    gcvSURF_L16F,
-    gcvSURF_A16L16F,
-    gcvSURF_A16R16F,
-    gcvSURF_A32F,
-    gcvSURF_L32F,
-    gcvSURF_A32L32F,
-    gcvSURF_A32R32F,
-    gcvSURF_E5B9G9R9,
-    gcvSURF_B10G11R11F,
-
-    gcvSURF_X16B16G16R16F_2_A8R8G8B8,
-    gcvSURF_A16B16G16R16F_2_A8R8G8B8,
-    gcvSURF_A16B16G16R16F_2_G16R16F,
-    gcvSURF_G32R32F_2_A8R8G8B8,
-    gcvSURF_X32B32G32R32F_2_G32R32F,
-    gcvSURF_A32B32G32R32F_2_G32R32F,
-    gcvSURF_X32B32G32R32F_4_A8R8G8B8,
-    gcvSURF_A32B32G32R32F_4_A8R8G8B8,
-
-    gcvSURF_R16F_1_A4R4G4B4,
-    gcvSURF_G16R16F_1_A8R8G8B8,
-    gcvSURF_B16G16R16F_2_A8R8G8B8,
-
-    gcvSURF_R32F_1_A8R8G8B8,
-    gcvSURF_B32G32R32F_3_A8R8G8B8,
-    gcvSURF_B10G11R11F_1_A8R8G8B8,
-
-    gcvSURF_A32F_1_R32F,
-    gcvSURF_L32F_1_R32F,
-    gcvSURF_A32L32F_1_G32R32F,
-
-
-
-    /* sRGB format. */
-    gcvSURF_SBGR8               = 1400,
-    gcvSURF_A8_SBGR8,
-    gcvSURF_X8_SBGR8,
-    gcvSURF_A8_SRGB8,
-    gcvSURF_X8_SRGB8,
-
-    /* Integer formats. */
-    gcvSURF_R8I                 = 1500,
-    gcvSURF_R8UI,
-    gcvSURF_R16I,
-    gcvSURF_R16UI,
-    gcvSURF_R32I,
-    gcvSURF_R32UI,
-    gcvSURF_X8R8I,
-    gcvSURF_G8R8I,
-    gcvSURF_X8R8UI,
-    gcvSURF_G8R8UI,
-    gcvSURF_X16R16I,
-    gcvSURF_G16R16I,
-    gcvSURF_X16R16UI,
-    gcvSURF_G16R16UI,
-    gcvSURF_X32R32I,
-    gcvSURF_G32R32I,
-    gcvSURF_X32R32UI,
-    gcvSURF_G32R32UI,
-    gcvSURF_X8G8R8I,
-    gcvSURF_B8G8R8I,
-    gcvSURF_X8G8R8UI,
-    gcvSURF_B8G8R8UI,
-    gcvSURF_X16G16R16I,
-    gcvSURF_B16G16R16I,
-    gcvSURF_X16G16R16UI,
-    gcvSURF_B16G16R16UI,
-    gcvSURF_X32G32R32I,
-    gcvSURF_B32G32R32I,
-    gcvSURF_X32G32R32UI,
-    gcvSURF_B32G32R32UI,
-    gcvSURF_X8B8G8R8I,
-    gcvSURF_A8B8G8R8I,
-    gcvSURF_X8B8G8R8UI,
-    gcvSURF_A8B8G8R8UI,
-    gcvSURF_X16B16G16R16I,
-    gcvSURF_A16B16G16R16I,
-    gcvSURF_X16B16G16R16UI,
-    gcvSURF_A16B16G16R16UI,
-    gcvSURF_X32B32G32R32I,
-    gcvSURF_A32B32G32R32I,
-    gcvSURF_X32B32G32R32UI,
-    gcvSURF_A32B32G32R32UI,
-    gcvSURF_A2B10G10R10UI,
-    gcvSURF_G32R32I_2_A8R8G8B8,
-    gcvSURF_G32R32I_1_G32R32F,
-    gcvSURF_G32R32UI_2_A8R8G8B8,
-    gcvSURF_G32R32UI_1_G32R32F,
-    gcvSURF_X16B16G16R16I_2_A8R8G8B8,
-    gcvSURF_X16B16G16R16I_1_G32R32F,
-    gcvSURF_A16B16G16R16I_2_A8R8G8B8,
-    gcvSURF_A16B16G16R16I_1_G32R32F,
-    gcvSURF_X16B16G16R16UI_2_A8R8G8B8,
-    gcvSURF_X16B16G16R16UI_1_G32R32F,
-    gcvSURF_A16B16G16R16UI_2_A8R8G8B8,
-    gcvSURF_A16B16G16R16UI_1_G32R32F,
-    gcvSURF_X32B32G32R32I_2_G32R32I,
-    gcvSURF_A32B32G32R32I_2_G32R32I,
-    gcvSURF_A32B32G32R32I_2_G32R32F,
-    gcvSURF_X32B32G32R32I_3_A8R8G8B8,
-    gcvSURF_A32B32G32R32I_4_A8R8G8B8,
-    gcvSURF_X32B32G32R32UI_2_G32R32UI,
-    gcvSURF_A32B32G32R32UI_2_G32R32UI,
-    gcvSURF_A32B32G32R32UI_2_G32R32F,
-    gcvSURF_X32B32G32R32UI_3_A8R8G8B8,
-    gcvSURF_A32B32G32R32UI_4_A8R8G8B8,
-    gcvSURF_A2B10G10R10UI_1_A8R8G8B8,
-    gcvSURF_A8B8G8R8I_1_A8R8G8B8,
-    gcvSURF_A8B8G8R8UI_1_A8R8G8B8,
-    gcvSURF_R8I_1_A4R4G4B4,
-    gcvSURF_R8UI_1_A4R4G4B4,
-    gcvSURF_R16I_1_A4R4G4B4,
-    gcvSURF_R16UI_1_A4R4G4B4,
-    gcvSURF_R32I_1_A8R8G8B8,
-    gcvSURF_R32UI_1_A8R8G8B8,
-    gcvSURF_X8R8I_1_A4R4G4B4,
-    gcvSURF_X8R8UI_1_A4R4G4B4,
-    gcvSURF_G8R8I_1_A4R4G4B4,
-    gcvSURF_G8R8UI_1_A4R4G4B4,
-    gcvSURF_X16R16I_1_A4R4G4B4,
-    gcvSURF_X16R16UI_1_A4R4G4B4,
-    gcvSURF_G16R16I_1_A8R8G8B8,
-    gcvSURF_G16R16UI_1_A8R8G8B8,
-    gcvSURF_X32R32I_1_A8R8G8B8,
-    gcvSURF_X32R32UI_1_A8R8G8B8,
-    gcvSURF_X8G8R8I_1_A4R4G4B4,
-    gcvSURF_X8G8R8UI_1_A4R4G4B4,
-    gcvSURF_B8G8R8I_1_A8R8G8B8,
-    gcvSURF_B8G8R8UI_1_A8R8G8B8,
-    gcvSURF_B16G16R16I_2_A8R8G8B8,
-    gcvSURF_B16G16R16I_1_G32R32F,
-    gcvSURF_B16G16R16UI_2_A8R8G8B8,
-    gcvSURF_B16G16R16UI_1_G32R32F,
-    gcvSURF_B32G32R32I_3_A8R8G8B8,
-    gcvSURF_B32G32R32UI_3_A8R8G8B8,
-    gcvSURF_A16B16G16R16_2_A8R8G8B8,
-    gcvSURF_R8G8B8_1_A8R8G8B8,
-    gcvSURF_G16R16_1_A8R8G8B8,
-    gcvSURF_A2B10G10R10_1_A8R8G8B8,
-    gcvSURF_A2R10G10B10_1_A8R8G8B8,
-    gcvSURF_A2W10V10U10_1_A8R8G8B8,
-
-    /* ASTC formats. */
-    gcvSURF_ASTC4x4             = 1600,
-    gcvSURF_ASTC5x4,
-    gcvSURF_ASTC5x5,
-    gcvSURF_ASTC6x5,
-    gcvSURF_ASTC6x6,
-    gcvSURF_ASTC8x5,
-    gcvSURF_ASTC8x6,
-    gcvSURF_ASTC8x8,
-    gcvSURF_ASTC10x5,
-    gcvSURF_ASTC10x6,
-    gcvSURF_ASTC10x8,
-    gcvSURF_ASTC10x10,
-    gcvSURF_ASTC12x10,
-    gcvSURF_ASTC12x12,
-    gcvSURF_ASTC4x4_SRGB,
-    gcvSURF_ASTC5x4_SRGB,
-    gcvSURF_ASTC5x5_SRGB,
-    gcvSURF_ASTC6x5_SRGB,
-    gcvSURF_ASTC6x6_SRGB,
-    gcvSURF_ASTC8x5_SRGB,
-    gcvSURF_ASTC8x6_SRGB,
-    gcvSURF_ASTC8x8_SRGB,
-    gcvSURF_ASTC10x5_SRGB,
-    gcvSURF_ASTC10x6_SRGB,
-    gcvSURF_ASTC10x8_SRGB,
-    gcvSURF_ASTC10x10_SRGB,
-    gcvSURF_ASTC12x10_SRGB,
-    gcvSURF_ASTC12x12_SRGB,
-
-    /* Recompile format*/
-    gcvSURF_L16_1_A4R4G4B4 = 1700,
-    gcvSURF_V16U16_1_A8R8G8B8,
-    gcvSURF_Q8W8V8U8_1_A8R8G8B8,
-    gcvSURF_X8L8V8U8_1_A8R8G8B8,
-    gcvSURF_R3G3B2_1_A8R8G8B8,
-    gcvSURF_A8R3G3B2_1_A8R8G8B8,
-    gcvSURF_W11V11U10_1_A8R8G8B8,
-    gcvSURF_Q16W16V16U16_2_A8R8G8B8,
-    gcvSURF_W11V11U10,
-    gcvSURF_V8U8_1_A4R4G4B4,
-    gcvSURF_A8B8G8R8_1_A8R8G8B8,
-    gcvSURF_A32R32G32B32_1_A8R8G8B8,
-    gcvSURF_X16B16G16R16F_1_A8R8G8B8,
-    gcvSURF_A16B16G16R16F_1_A8R8G8B8,
-    gcvSURF_G32R32F_1_A8R8G8B8,
-    gcvSURF_X32B32G32R32F_1_A8R8G8B8,
-    gcvSURF_A32B32G32R32F_1_A8R8G8B8,
-    gcvSURF_G32R32I_1_A8R8G8B8,
-    gcvSURF_G32R32UI_1_A8R8G8B8,
-    gcvSURF_A32B32G32R32I_1_A8R8G8B8,
-    gcvSURF_A32B32G32R32UI_1_A8R8G8B8,
-    gcvSURF_Q16W16V16U16_1_A8R8G8B8,
-    gcvSURF_A16B16G16R16_1_A8R8G8B8,
-
-    /* GL4 formats */
-    gcvSURF_G8                  = 1800,
-    gcvSURF_B8,
-    gcvSURF_G32F
-}
-gceSURF_FORMAT;
 
 typedef enum _gceIMAGE_MEM_TYPE
 {
@@ -1812,90 +683,11 @@ typedef enum _gce2D_SOURCE
 }
 gce2D_SOURCE;
 
-/* Pipes. */
-typedef enum _gcePIPE_SELECT
-{
-    gcvPIPE_INVALID = ~0,
-    gcvPIPE_3D      =  0,
-    gcvPIPE_2D
-}
-gcePIPE_SELECT;
-
-/* Hardware type. */
-typedef enum _gceHARDWARE_TYPE
-{
-    gcvHARDWARE_INVALID,
-    gcvHARDWARE_3D2D,
-    gcvHARDWARE_3D,
-    gcvHARDWARE_2D,
-    gcvHARDWARE_VIP,
-    gcvHARDWARE_VG,
-    gcvHARDWARE_NUM_TYPES,
-}
-gceHARDWARE_TYPE;
-
-#define gcdCHIP_COUNT               gcvCORE_COUNT
-
 typedef enum _gceMMU_MODE
 {
     gcvMMU_MODE_1K,
     gcvMMU_MODE_4K,
 } gceMMU_MODE;
-
-/* User signal command codes. */
-typedef enum _gceUSER_SIGNAL_COMMAND_CODES
-{
-    gcvUSER_SIGNAL_CREATE,
-    gcvUSER_SIGNAL_DESTROY,
-    gcvUSER_SIGNAL_SIGNAL,
-    gcvUSER_SIGNAL_WAIT,
-    gcvUSER_SIGNAL_MAP,
-    gcvUSER_SIGNAL_UNMAP,
-}
-gceUSER_SIGNAL_COMMAND_CODES;
-
-/* Shared buffer command codes. */
-typedef enum _gceSHBUF_COMMAND_CODES
-{
-    gcvSHBUF_CREATE,
-    gcvSHBUF_DESTROY,
-    gcvSHBUF_MAP,
-    gcvSHBUF_WRITE,
-    gcvSHBUF_READ,
-}
-gceSHBUF_COMMAND_CODES;
-
-/* Event locations. */
-typedef enum _gceKERNEL_WHERE
-{
-    gcvKERNEL_COMMAND,
-    gcvKERNEL_VERTEX,
-    gcvKERNEL_TRIANGLE,
-    gcvKERNEL_TEXTURE,
-    gcvKERNEL_PIXEL,
-    gcvKERNEL_BLT,
-}
-gceKERNEL_WHERE;
-
-#if gcdENABLE_VG
-/* Hardware blocks. */
-typedef enum _gceBLOCK
-{
-    gcvBLOCK_COMMAND,
-    gcvBLOCK_TESSELLATOR,
-    gcvBLOCK_TESSELLATOR2,
-    gcvBLOCK_TESSELLATOR3,
-    gcvBLOCK_RASTER,
-    gcvBLOCK_VG,
-    gcvBLOCK_VG2,
-    gcvBLOCK_VG3,
-    gcvBLOCK_PIXEL,
-
-    /* Number of defined blocks. */
-    gcvBLOCK_COUNT
-}
-gceBLOCK;
-#endif
 
 /* gcdDUMP message type. */
 typedef enum _gceDEBUG_MESSAGE_TYPE
@@ -2108,24 +900,6 @@ typedef enum _gceMULTI_GPU_RENDERING_MODE
 }
 gceMULTI_GPU_RENDERING_MODE;
 
-typedef enum _gceCORE_3D_MASK
-{
-    gcvCORE_3D_0_MASK   = (1 << 0),
-    gcvCORE_3D_1_MASK   = (1 << 1),
-
-    gcvCORE_3D_ALL_MASK = (0xFFFF)
-}
-gceCORE_3D_MASK;
-
-typedef enum _gceCORE_3D_ID
-{
-    gcvCORE_3D_0_ID       = 0,
-    gcvCORE_3D_1_ID       = 1,
-
-    gcvCORE_3D_ID_INVALID = ~0UL
-}
-gceCORE_3D_ID;
-
 typedef enum _gceMULTI_GPU_MODE
 {
     gcvMULTI_GPU_MODE_COMBINED    = 0,
@@ -2162,85 +936,6 @@ typedef enum _gceHAL_ARG_VERSION
 gceHAL_ARG_VERSION;
 
 
-typedef enum _gceCHIP_FLAG
-{
-    gcvCHIP_FLAG_MSAA_COHERENCEY_ECO_FIX = 1 << 0,
-    gcvCHIP_FLAG_GC2000_R2               = 1 << 1,
-    gcvCHIP_AXI_BUS128_BITS              = 1 << 2,
-}
-gceCHIP_FLAG;
-
-/* If different, choose render engine */
-#define PRIORITY_ENGINE(a, b) gcmMIN(a,b)
-
-typedef enum
-{
-    gcvENGINE_RENDER            = 0,
-    gcvENGINE_BLT               = 1,
-    gcvENGINE_GPU_ENGINE_COUNT  = 2,
-    gcvENGINE_CPU               = gcvENGINE_GPU_ENGINE_COUNT,
-    gcvENGINE_ALL_COUNT         = gcvENGINE_CPU + 1,
-    gcvENGINE_INVALID           = gcvENGINE_ALL_COUNT + 0x100
-}
-gceENGINE;
-
-/* CORE enum. */
-typedef enum _gceCORE
-{
-    gcvCORE_MAJOR,
-    gcvCORE_3D1,
-    gcvCORE_3D2,
-    gcvCORE_3D3,
-    gcvCORE_3D4,
-    gcvCORE_3D5,
-    gcvCORE_3D6,
-    gcvCORE_3D7,
-    gcvCORE_3D_MAX = gcvCORE_3D7,
-    gcvCORE_2D,
-    gcvCORE_VG,
-#if gcdDEC_ENABLE_AHB
-    gcvCORE_DEC,
-#endif
-    gcvCORE_COUNT
-}
-gceCORE;
-
-
-typedef enum _gceSECURE_MODE
-{
-    /* For cores without gcvFEATURE_SECURITY. */
-    gcvSECURE_NONE,
-
-    /* Use registers added in gcvFEATURE_SECURITY in normal driver,
-    ** In this mode, GPU always works under non secure mode and
-    ** should not touch secure buffer. It is used to test basic function.
-    */
-    gcvSECURE_IN_NORMAL,
-
-    /* Make use of gcvFEATURE_SECURITY in trust application. */
-    gcvSECURE_IN_TA
-}
-gceSECURE_MODE;
-
-/* kernel driver compression option, as it's a system global option,
-** it means kernel driver allows the options, NOT necessarily means it must be on.
-*/
-typedef enum _gceCOMPRESSION_OPTION
-{
-    gcvCOMPRESSION_OPTION_NONE       = 0x0, /* No any compression */
-    gcvCOMPRESSION_OPTION_COLOR      = 0x1, /* Compression for non-msaa color format */
-    gcvCOMPRESSION_OPTION_DEPTH      = 0x2, /* Compression for non-msaa depth format */
-    gcvCOMPRESSION_OPTION_MSAA_COLOR = 0x4, /* Compression for msaa color */
-    gcvCOMPRESSION_OPTION_MSAA_DEPTH = 0x8, /* Compression for msaa depth */
-
-    /* default compressio option */
-    gcvCOMPRESSION_OPTION_DEFAULT    = gcvCOMPRESSION_OPTION_DEPTH      |
-                                       gcvCOMPRESSION_OPTION_COLOR      |
-                                       gcvCOMPRESSION_OPTION_MSAA_COLOR |
-                                       gcvCOMPRESSION_OPTION_MSAA_DEPTH,
-}
-gceCOMPRESSION_OPTION;
-
 typedef enum _gceHW_FE_TYPE
 {
     gcvHW_FE_WAIT_LINK,
@@ -2261,31 +956,6 @@ typedef enum _gceMCFE_CHANNEL_TYPE
 }
 gceMCFE_CHANNEL_TYPE;
 
-typedef enum _gceSRAM_INTERNAL
-{
-    gcvSRAM_INTERNAL0 = 0,
-    gcvSRAM_INTERNAL1,
-
-    gcvSRAM_INTER_COUNT
-}
-gceSRAM_INTERNAL;
-
-typedef enum _gceSRAM_EXTERNAL
-{
-    gcvSRAM_EXTERNAL0 = 0,
-    gcvSRAM_EXTERNAL1,
-
-    gcvSRAM_EXT_COUNT
-}
-gceSRAM_EXTERNAL;
-
-typedef enum _gceFLATMAP_FLAG
-{
-    gcvFLATMAP_DIRECT,
-    gcvFLATMAP_SHIFT,
-}
-gceFLATMAP_FLAG;
-
 typedef enum _gcePAGE_TYPE
 {
     gcvPAGE_TYPE_1M,
@@ -2302,29 +972,1104 @@ typedef enum _gceAREA_TYPE
 }
 gceAREA_TYPE;
 
-/* Video memory alloation type. */
-typedef enum _gceVIDMEM_TYPE
+/*******************************************************************************
+** Broadcast interface.
+*/
+
+typedef enum _gceBROADCAST
 {
-    gcvVIDMEM_TYPE_GENERIC          = gcvSURF_TYPE_UNKNOWN,
-    gcvVIDMEM_TYPE_INDEX_BUFFER     = gcvSURF_INDEX,
-    gcvVIDMEM_TYPE_VERTEX_BUFFER    = gcvSURF_VERTEX,
-    gcvVIDMEM_TYPE_TEXTURE          = gcvSURF_TEXTURE,
-    gcvVIDMEM_TYPE_COLOR_BUFFER     = gcvSURF_RENDER_TARGET,
-    gcvVIDMEM_TYPE_DEPTH_BUFFER     = gcvSURF_DEPTH,
-    gcvVIDMEM_TYPE_BITMAP           = gcvSURF_BITMAP,
-    gcvVIDMEM_TYPE_TILE_STATUS      = gcvSURF_TILE_STATUS,
-    gcvVIDMEM_TYPE_IMAGE            = gcvSURF_IMAGE,
-    gcvVIDMEM_TYPE_MASK             = gcvSURF_MASK,
-    gcvVIDMEM_TYPE_SCISSOR          = gcvSURF_SCISSOR,
-    gcvVIDMEM_TYPE_HZ_BUFFER        = gcvSURF_HIERARCHICAL_DEPTH,
-    gcvVIDMEM_TYPE_ICACHE           = gcvSURF_ICACHE,
-    gcvVIDMEM_TYPE_TXDESC           = gcvSURF_TXDESC,
-    gcvVIDMEM_TYPE_FENCE            = gcvSURF_FENCE,
-    gcvVIDMEM_TYPE_TFBHEADER        = gcvSURF_TFBHEADER,
-    gcvVIDMEM_TYPE_COMMAND,
-    gcvVIDMEM_TYPE_COUNT
+    /* GPU might be idle. */
+    gcvBROADCAST_GPU_IDLE,
+
+    /* A commit is going to happen. */
+    gcvBROADCAST_GPU_COMMIT,
+
+    /* GPU seems to be stuck. */
+    gcvBROADCAST_GPU_STUCK,
+
+    /* First process gets attached. */
+    gcvBROADCAST_FIRST_PROCESS,
+
+    /* Last process gets detached. */
+    gcvBROADCAST_LAST_PROCESS,
+
+    /* AXI bus error. */
+    gcvBROADCAST_AXI_BUS_ERROR,
+
+    /* Out of memory. */
+    gcvBROADCAST_OUT_OF_MEMORY,
 }
-gceVIDMEM_TYPE;
+gceBROADCAST;
+
+/* Notifications. */
+typedef enum _gceNOTIFY
+{
+    gcvNOTIFY_INTERRUPT,
+    gcvNOTIFY_COMMAND_QUEUE,
+}
+gceNOTIFY;
+
+/* Flush flags. */
+typedef enum _gceKERNEL_FLUSH
+{
+    gcvFLUSH_COLOR              = 0x01,
+    gcvFLUSH_DEPTH              = 0x02,
+    gcvFLUSH_TEXTURE            = 0x04,
+    gcvFLUSH_2D                 = 0x08,
+    gcvFLUSH_L2                 = 0x10,
+    gcvFLUSH_TILE_STATUS        = 0x20,
+    gcvFLUSH_ICACHE             = 0x40,
+    gcvFLUSH_TXDESC             = 0x80,
+    gcvFLUSH_FENCE              = 0x100,
+    gcvFLUSH_VERTEX             = 0x200,
+    gcvFLUSH_TFBHEADER          = 0x400,
+    gcvFLUSH_ALL                = gcvFLUSH_COLOR
+                                | gcvFLUSH_DEPTH
+                                | gcvFLUSH_TEXTURE
+                                | gcvFLUSH_2D
+                                | gcvFLUSH_L2
+                                | gcvFLUSH_TILE_STATUS
+                                | gcvFLUSH_ICACHE
+                                | gcvFLUSH_TXDESC
+                                | gcvFLUSH_FENCE
+                                | gcvFLUSH_VERTEX
+                                | gcvFLUSH_TFBHEADER
+}
+gceKERNEL_FLUSH;
+
+typedef enum _gceCOUNTER
+{
+    gcvCOUNTER_FRONT_END,
+    gcvCOUNTER_VERTEX_SHADER,
+    gcvCOUNTER_PRIMITIVE_ASSEMBLY,
+    gcvCOUNTER_SETUP,
+    gcvCOUNTER_RASTERIZER,
+    gcvCOUNTER_PIXEL_SHADER,
+    gcvCOUNTER_TEXTURE,
+    gcvCOUNTER_PIXEL_ENGINE,
+    gcvCOUNTER_MEMORY_CONTROLLER_COLOR,
+    gcvCOUNTER_MEMORY_CONTROLLER_DEPTH,
+    gcvCOUNTER_HOST_INTERFACE0,
+    gcvCOUNTER_HOST_INTERFACE1,
+    gcvCOUNTER_GPUL2_CACHE,
+    gcvCOUNTER_COUNT
+}
+gceCOUNTER;
+
+typedef enum _gceProfilerClient
+{
+    gcvCLIENT_OPENGLES11 = 1,
+    gcvCLIENT_OPENGLES,
+    gcvCLIENT_OPENGL,
+    gcvCLIENT_OPENVG,
+    gcvCLIENT_OPENCL,
+    gcvCLIENT_OPENVX,
+    gcvCLIENT_OPENVK,
+}
+gceProfilerClient;
+
+typedef enum _gceCOUNTER_OPTYPE
+{
+    gcvCOUNTER_OP_DRAW = 0,
+    gcvCOUNTER_OP_BLT = 1,
+    gcvCOUNTER_OP_COMPUTE = 2,
+    gcvCOUNTER_OP_RS = 3,
+    gcvCOUNTER_OP_FINISH = 4,
+    gcvCOUNTER_OP_FRAME = 5,
+    gcvCOUNTER_OP_NONE = 6
+}
+gceCOUNTER_OPTYPE;
+
+typedef enum _gceProbeStatus
+{
+    gcvPROBE_Disabled = 0,
+    gcvPROBE_Paused = 1,
+    gcvPROBE_Enabled = 2,
+}
+gceProbeStatus;
+
+typedef enum _gceProbeCmd
+{
+    gcvPROBECMD_BEGIN = 0,
+    gcvPROBECMD_PAUSE = 1,
+    gcvPROBECMD_RESUME = 2,
+    gcvPROBECMD_END = 3,
+}
+gceProbeCmd;
+
+/*******************************************************************************
+** Events. *********************************************************************
+*/
+
+typedef enum _halEventType
+{
+    /* Keyboard event. */
+    HAL_KEYBOARD,
+
+    /* Mouse move event. */
+    HAL_POINTER,
+
+    /* Mouse button event. */
+    HAL_BUTTON,
+
+    /* Application close event. */
+    HAL_CLOSE,
+
+    /* Application window has been updated. */
+    HAL_WINDOW_UPDATE
+}
+halEventType;
+
+/* Scancodes for keyboard. */
+typedef enum _halKeys
+{
+    HAL_UNKNOWN = -1,
+
+    HAL_BACKSPACE = 0x08,
+    HAL_TAB,
+    HAL_ENTER = 0x0D,
+    HAL_ESCAPE = 0x1B,
+
+    HAL_SPACE = 0x20,
+    HAL_SINGLEQUOTE = 0x27,
+    HAL_PAD_ASTERISK = 0x2A,
+    HAL_COMMA = 0x2C,
+    HAL_HYPHEN,
+    HAL_PERIOD,
+    HAL_SLASH,
+    HAL_0,
+    HAL_1,
+    HAL_2,
+    HAL_3,
+    HAL_4,
+    HAL_5,
+    HAL_6,
+    HAL_7,
+    HAL_8,
+    HAL_9,
+    HAL_SEMICOLON = 0x3B,
+    HAL_EQUAL = 0x3D,
+    HAL_A = 0x41,
+    HAL_B,
+    HAL_C,
+    HAL_D,
+    HAL_E,
+    HAL_F,
+    HAL_G,
+    HAL_H,
+    HAL_I,
+    HAL_J,
+    HAL_K,
+    HAL_L,
+    HAL_M,
+    HAL_N,
+    HAL_O,
+    HAL_P,
+    HAL_Q,
+    HAL_R,
+    HAL_S,
+    HAL_T,
+    HAL_U,
+    HAL_V,
+    HAL_W,
+    HAL_X,
+    HAL_Y,
+    HAL_Z,
+    HAL_LBRACKET,
+    HAL_BACKSLASH,
+    HAL_RBRACKET,
+    HAL_BACKQUOTE = 0x60,
+
+    HAL_F1 = 0x80,
+    HAL_F2,
+    HAL_F3,
+    HAL_F4,
+    HAL_F5,
+    HAL_F6,
+    HAL_F7,
+    HAL_F8,
+    HAL_F9,
+    HAL_F10,
+    HAL_F11,
+    HAL_F12,
+
+    HAL_LCTRL,
+    HAL_RCTRL,
+    HAL_LSHIFT,
+    HAL_RSHIFT,
+    HAL_LALT,
+    HAL_RALT,
+    HAL_CAPSLOCK,
+    HAL_NUMLOCK,
+    HAL_SCROLLLOCK,
+    HAL_PAD_0,
+    HAL_PAD_1,
+    HAL_PAD_2,
+    HAL_PAD_3,
+    HAL_PAD_4,
+    HAL_PAD_5,
+    HAL_PAD_6,
+    HAL_PAD_7,
+    HAL_PAD_8,
+    HAL_PAD_9,
+    HAL_PAD_HYPHEN,
+    HAL_PAD_PLUS,
+    HAL_PAD_SLASH,
+    HAL_PAD_PERIOD,
+    HAL_PAD_ENTER,
+    HAL_SYSRQ,
+    HAL_PRNTSCRN,
+    HAL_BREAK,
+    HAL_UP,
+    HAL_LEFT,
+    HAL_RIGHT,
+    HAL_DOWN,
+    HAL_HOME,
+    HAL_END,
+    HAL_PGUP,
+    HAL_PGDN,
+    HAL_INSERT,
+    HAL_DELETE,
+    HAL_LWINDOW,
+    HAL_RWINDOW,
+    HAL_MENU,
+    HAL_POWER,
+    HAL_SLEEP,
+    HAL_WAKE
+}
+halKeys;
+
+/*!
+ @brief Command codes between kernel module and TrustZone
+ @discussion
+ Critical services must be done in TrustZone to avoid sensitive content leak. Most of kernel module is kept in non-Secure os to minimize
+ code in TrustZone.
+ */
+typedef enum kernel_packet_command {
+    KERNEL_START_COMMAND,
+    KERNEL_SUBMIT,
+    KERNEL_MAP_MEMORY, /* */
+    KERNEL_UNMAP_MEMORY,
+    KERNEL_ALLOCATE_SECRUE_MEMORY, /*! Security memory management. */
+    KERNEL_FREE_SECURE_MEMORY,
+    KERNEL_EXECUTE, /* Execute a command buffer. */
+    KERNEL_DUMP_MMU_EXCEPTION,
+    KERNEL_HANDLE_MMU_EXCEPTION,
+    KERNEL_READ_MMU_EXCEPTION,
+} kernel_packet_command_t;
+
+enum {
+    gcvTA_COMMAND_INIT,
+    gcvTA_COMMAND_DISPATCH,
+
+    gcvTA_CALLBACK_ALLOC_SECURE_MEM,
+    gcvTA_CALLBACK_FREE_SECURE_MEM,
+};
+
+typedef enum {
+    gcvFENCE_TYPE_READ          = 0x1,
+    gcvFENCE_TYPE_WRITE         = 0x2,
+    gcvFENCE_TYPE_ALL           = gcvFENCE_TYPE_READ | gcvFENCE_TYPE_WRITE,
+    gcvFNECE_TYPE_INVALID       = 0x10000,
+}
+gceFENCE_TYPE;
+
+typedef enum _gceTLS_KEY
+{
+    gcvTLS_KEY_EGL,
+    gcvTLS_KEY_OPENGL_ES,
+    gcvTLS_KEY_OPENVG,
+    gcvTLS_KEY_OPENGL,
+    gcvTLS_KEY_OPENCL,
+    gcvTLS_KEY_OPENVX,
+
+    gcvTLS_KEY_COUNT
+}
+gceTLS_KEY;
+
+typedef enum _gcePLS_VALUE
+{
+  gcePLS_VALUE_EGL_DISPLAY_INFO,
+  gcePLS_VALUE_EGL_CONFIG_FORMAT_INFO,
+  gcePLS_VALUE_EGL_DESTRUCTOR_INFO,
+}
+gcePLS_VALUE;
+
+#if gcdENABLE_3D
+/* Blending functions. */
+typedef enum _gceBLEND_FUNCTION
+{
+    gcvBLEND_ZERO,
+    gcvBLEND_ONE,
+    gcvBLEND_SOURCE_COLOR,
+    gcvBLEND_INV_SOURCE_COLOR,
+    gcvBLEND_SOURCE_ALPHA,
+    gcvBLEND_INV_SOURCE_ALPHA,
+    gcvBLEND_TARGET_COLOR,
+    gcvBLEND_INV_TARGET_COLOR,
+    gcvBLEND_TARGET_ALPHA,
+    gcvBLEND_INV_TARGET_ALPHA,
+    gcvBLEND_SOURCE_ALPHA_SATURATE,
+    gcvBLEND_CONST_COLOR,
+    gcvBLEND_INV_CONST_COLOR,
+    gcvBLEND_CONST_ALPHA,
+    gcvBLEND_INV_CONST_ALPHA,
+}
+gceBLEND_FUNCTION;
+
+/* Blending modes. */
+typedef enum _gceBLEND_MODE
+{
+    gcvBLEND_ADD = 0,
+    gcvBLEND_SUBTRACT,
+    gcvBLEND_REVERSE_SUBTRACT,
+    gcvBLEND_MIN,
+    gcvBLEND_MAX,
+    gcvBLEND_MULTIPLY,
+    gcvBLEND_SCREEN,
+    gcvBLEND_OVERLAY,
+    gcvBLEND_DARKEN,
+    gcvBLEND_LIGHTEN,
+    gcvBLEND_COLORDODGE,
+    gcvBLEND_COLORBURN,
+    gcvBLEND_HARDLIGHT,
+    gcvBLEND_SOFTLIGHT,
+    gcvBLEND_DIFFERENCE,
+    gcvBLEND_EXCLUSION,
+    gcvBLEND_HSL_HUE,
+    gcvBLEND_HSL_SATURATION,
+    gcvBLEND_HSL_COLOR,
+    gcvBLEND_HSL_LUMINOSITY,
+
+    gcvBLEND_TOTAL
+}
+gceBLEND_MODE;
+
+/* Depth modes. */
+typedef enum _gceDEPTH_MODE
+{
+    gcvDEPTH_NONE,
+    gcvDEPTH_Z,
+    gcvDEPTH_W,
+}
+gceDEPTH_MODE;
+#endif /* gcdENABLE_3D */
+
+/* API flags. */
+typedef enum _gceAPI
+{
+    gcvAPI_D3D = 1,
+    gcvAPI_OPENGL_ES11,
+    gcvAPI_OPENGL_ES20,
+    gcvAPI_OPENGL_ES30,
+    gcvAPI_OPENGL_ES31,
+    gcvAPI_OPENGL_ES32,
+    gcvAPI_OPENGL,
+    gcvAPI_OPENVG,
+    gcvAPI_OPENCL,
+    gcvAPI_OPENVK,
+}
+gceAPI;
+
+typedef enum _gceWHERE
+{
+    gcvWHERE_COMMAND_PREFETCH = 0,
+    gcvWHERE_COMMAND,
+    gcvWHERE_RASTER,
+    gcvWHERE_PIXEL,
+    gcvWHERE_BLT,
+}
+gceWHERE;
+
+typedef enum _gceHOW
+{
+    gcvHOW_SEMAPHORE            = 0x1,
+    gcvHOW_STALL                = 0x2,
+    gcvHOW_SEMAPHORE_STALL      = 0x3,
+}
+gceHOW;
+
+typedef enum _gceSignalHandlerType
+{
+    gcvHANDLE_SIGFPE_WHEN_SIGNAL_CODE_IS_0        = 0x1,
+}
+gceSignalHandlerType;
+
+typedef enum _gceFILE_MODE
+{
+    gcvFILE_CREATE          = 0,
+    gcvFILE_APPEND,
+    gcvFILE_READ,
+    gcvFILE_CREATETEXT,
+    gcvFILE_APPENDTEXT,
+    gcvFILE_READTEXT,
+}
+gceFILE_MODE;
+
+typedef enum _gceFILE_WHENCE
+{
+    gcvFILE_SEEK_SET,
+    gcvFILE_SEEK_CUR,
+    gcvFILE_SEEK_END
+}
+gceFILE_WHENCE;
+
+/* Color format classes. */
+typedef enum _gceFORMAT_CLASS
+{
+    gcvFORMAT_CLASS_RGBA        = 4500,
+    gcvFORMAT_CLASS_YUV,
+    gcvFORMAT_CLASS_INDEX,
+    gcvFORMAT_CLASS_LUMINANCE,
+    gcvFORMAT_CLASS_BUMP,
+    gcvFORMAT_CLASS_DEPTH,
+    gcvFORMAT_CLASS_ASTC,
+    gcvFORMAT_CLASS_COMPRESSED,
+    gcvFORMAT_CLASS_OTHER
+}
+gceFORMAT_CLASS;
+
+/* Color format data type */
+typedef enum _gceFORMAT_DATATYPE
+{
+    gcvFORMAT_DATATYPE_UNSIGNED_NORMALIZED,
+    gcvFORMAT_DATATYPE_SIGNED_NORMALIZED,
+    gcvFORMAT_DATATYPE_UNSIGNED_INTEGER,
+    gcvFORMAT_DATATYPE_SIGNED_INTEGER,
+    gcvFORMAT_DATATYPE_FLOAT16,
+    gcvFORMAT_DATATYPE_FLOAT32,
+    gcvFORMAT_DATATYPE_FLOAT_E5B9G9R9,
+    gcvFORMAT_DATATYPE_FLOAT_B10G11R11F,
+    gcvFORMAT_DATATYPE_INDEX,
+    gcvFORMAT_DATATYPE_SRGB,
+    gcvFORMAT_DATATYPE_FLOAT32_UINT,
+}
+gceFORMAT_DATATYPE;
+
+typedef enum _gceORIENTATION
+{
+    gcvORIENTATION_TOP_BOTTOM,
+    gcvORIENTATION_BOTTOM_TOP,
+}
+gceORIENTATION;
+
+/* Special enums for width field in gcsFORMAT_COMPONENT. */
+typedef enum _gceCOMPONENT_CONTROL
+{
+    gcvCOMPONENT_NOTPRESENT     = 0x00,
+    gcvCOMPONENT_DONTCARE       = 0x80,
+    gcvCOMPONENT_WIDTHMASK      = 0x7F,
+    gcvCOMPONENT_ODD            = 0x80
+}
+gceCOMPONENT_CONTROL;
+
+/* User option. */
+typedef enum _gceDEBUG_MSG
+{
+    gcvDEBUG_MSG_NONE,
+    gcvDEBUG_MSG_ERROR,
+    gcvDEBUG_MSG_WARNING
+}
+gceDEBUG_MSG;
+
+/* Compressed format now was defined same as dec400d, should be general. */
+typedef enum _VIV_COMPRESS_FMT
+{
+    _VIV_CFMT_ARGB8 = 0,
+    _VIV_CFMT_XRGB8,
+    _VIV_CFMT_AYUV,
+    _VIV_CFMT_UYVY,
+    _VIV_CFMT_YUY2,
+    _VIV_CFMT_YUV_ONLY,
+    _VIV_CFMT_UV_MIX,
+    _VIV_CFMT_ARGB4,
+    _VIV_CFMT_XRGB4,
+    _VIV_CFMT_A1R5G5B5,
+    _VIV_CFMT_X1R5G5B5,
+    _VIV_CFMT_R5G6B5,
+    _VIV_CFMT_Z24S8,
+    _VIV_CFMT_Z24,
+    _VIV_CFMT_Z16,
+    _VIV_CFMT_A2R10G10B10,
+    _VIV_CFMT_BAYER,
+    _VIV_CFMT_SIGNED_BAYER,
+    _VIV_CFMT_VAA16,
+    _VIV_CFMT_S8,
+
+    _VIV_CFMT_MAX,
+} _VIV_COMPRESS_FMT;
+
+typedef enum _gcePROGRAM_STAGE
+{
+    gcvPROGRAM_STAGE_VERTEX         = 0x0,
+    gcvPROGRAM_STAGE_TCS            = 0x1,
+    gcvPROGRAM_STAGE_TES            = 0x2,
+    gcvPROGRAM_STAGE_GEOMETRY       = 0x3,
+    gcvPROGRAM_STAGE_FRAGMENT       = 0x4,
+    gcvPROGRAM_STAGE_GRAPHICS_COUNT = 0x5,
+    gcvPROGRAM_STAGE_COMPUTE        = 0x5,
+    gcvPROGRAM_STAGE_OPENCL         = 0x6,
+    gcvPROGRAM_STAGE_LAST
+}
+gcePROGRAM_STAGE;
+
+typedef enum _gcePROGRAM_STAGE_BIT
+{
+    gcvPROGRAM_STAGE_VERTEX_BIT   = 1 << gcvPROGRAM_STAGE_VERTEX,
+    gcvPROGRAM_STAGE_TCS_BIT      = 1 << gcvPROGRAM_STAGE_TCS,
+    gcvPROGRAM_STAGE_TES_BIT      = 1 << gcvPROGRAM_STAGE_TES,
+    gcvPROGRAM_STAGE_GEOMETRY_BIT = 1 << gcvPROGRAM_STAGE_GEOMETRY,
+    gcvPROGRAM_STAGE_FRAGMENT_BIT = 1 << gcvPROGRAM_STAGE_FRAGMENT,
+    gcvPROGRAM_STAGE_COMPUTE_BIT  = 1 << gcvPROGRAM_STAGE_COMPUTE,
+    gcvPROGRAM_STAGE_OPENCL_BIT   = 1 << gcvPROGRAM_STAGE_OPENCL,
+}
+gcePROGRAM_STAGE_BIT;
+
+typedef enum _gceBLIT_FLAG
+{
+    gcvBLIT_FLAG_SKIP_DEPTH_WRITE   = 1 << 0,
+    gcvBLIT_FLAG_SKIP_STENCIL_WRITE = 1 << 1,
+} gceBLIT_FLAG;
+
+/* Clear flags. */
+typedef enum _gceCLEAR
+{
+    gcvCLEAR_COLOR              = 0x1,
+    gcvCLEAR_DEPTH              = 0x2,
+    gcvCLEAR_STENCIL            = 0x4,
+    gcvCLEAR_HZ                 = 0x8,
+    gcvCLEAR_WITH_GPU_ONLY      = 0x100,
+    gcvCLEAR_WITH_CPU_ONLY      = 0x200,
+    gcvCLEAR_MULTI_SLICES       = 0x400,
+}
+gceCLEAR;
+
+typedef enum _gceBLITDRAW_TYPE
+{
+    gcvBLITDRAW_CLEAR = 0,
+    gcvBLITDRAW_BLIT  = 1,
+
+    /* last number, not a real type */
+    gcvBLITDRAW_NUM_TYPE
+ }
+gceBLITDRAW_TYPE;
+
+typedef enum _gceSPLIT_DRAW_TYPE
+{
+    gcvSPLIT_DRAW_UNKNOWN      = 0x0,
+    gcvSPLIT_DRAW_1,
+    gcvSPLIT_DRAW_2,
+    gcvSPLIT_DRAW_3,
+    gcvSPLIT_DRAW_4,
+    gcvSPLIT_DRAW_XFB,
+    gcvSPLIT_DRAW_INDEX_FETCH,
+    gcvSPLIT_DRAW_TCS,
+    gcvSPLIT_DRAW_STIPPLE,
+    gcvSPLIT_DRAW_WIDE_LINE,
+    gcvSPLIT_DRAW_LAST
+}
+gceSPLIT_DRAW_TYPE;
+
+/* Blending targets. */
+typedef enum _gceBLEND_UNIT
+{
+    gcvBLEND_SOURCE,
+    gcvBLEND_TARGET,
+}
+gceBLEND_UNIT;
+
+typedef enum _gceXfbCmd
+{
+    gcvXFBCMD_BEGIN           = 0,
+    gcvXFBCMD_PAUSE           = 1,
+    gcvXFBCMD_RESUME          = 2,
+    gcvXFBCMD_END             = 3,
+    gcvXFBCMD_PAUSE_INCOMMIT  = 4,
+    gcvXFBCMD_RESUME_INCOMMIT = 5,
+    gcvXFBCMD_INVALID         = 6,
+}
+gceXfbCmd;
+
+typedef enum _gceXfbStatus
+{
+    gcvXFB_Disabled = 0,
+    gcvXFB_Paused,
+    gcvXFB_Enabled,
+}
+gceXfbStatus;
+
+typedef enum _gceQueryStatus
+{
+    gcvQUERY_Disabled = 0,
+    gcvQUERY_Paused   = 1,
+    gcvQUERY_Enabled  = 2,
+}
+gceQueryStatus;
+
+typedef enum _gceQueryCmd
+{
+    gcvQUERYCMD_BEGIN   = 0,
+    gcvQUERYCMD_PAUSE   = 1,
+    gcvQUERYCMD_RESUME  = 2,
+    gcvQUERYCMD_END     = 3,
+    gcvQUERYCMD_INVALID = 4,
+}
+gceQueryCmd;
+
+typedef enum _gceQueryType
+{
+    gcvQUERY_OCCLUSION = 0,
+    gcvQUERY_XFB_WRITTEN = 1,
+    gcvQUERY_PRIM_GENERATED = 2,
+    gcvQUERY_MAX_NUM = 3,
+}
+gceQueryType;
+
+/* Cube faces. */
+typedef enum _gceTEXTURE_FACE
+{
+    gcvFACE_NONE = 0,
+    gcvFACE_POSITIVE_X,
+    gcvFACE_NEGATIVE_X,
+    gcvFACE_POSITIVE_Y,
+    gcvFACE_NEGATIVE_Y,
+    gcvFACE_POSITIVE_Z,
+    gcvFACE_NEGATIVE_Z,
+}
+gceTEXTURE_FACE;
+
+typedef enum _gceVERTEX_FORMAT
+{
+    gcvVERTEX_BYTE,
+    gcvVERTEX_UNSIGNED_BYTE,
+    gcvVERTEX_SHORT,
+    gcvVERTEX_UNSIGNED_SHORT,
+    gcvVERTEX_INT,
+    gcvVERTEX_UNSIGNED_INT,
+    gcvVERTEX_FIXED,
+    gcvVERTEX_HALF,
+    gcvVERTEX_FLOAT,
+    gcvVERTEX_DOUBLE,
+    gcvVERTEX_UNSIGNED_INT_10_10_10_2,
+    gcvVERTEX_INT_10_10_10_2,
+    gcvVERTEX_UNSIGNED_INT_2_10_10_10_REV,
+    gcvVERTEX_INT_2_10_10_10_REV,
+    /* integer format */
+    gcvVERTEX_INT8,
+    gcvVERTEX_INT16,
+    gcvVERTEX_INT32,
+}
+gceVERTEX_FORMAT;
+
+/* What the SW converting scheme to create temp attrib */
+typedef enum _gceATTRIB_SCHEME
+{
+    gcvATTRIB_SCHEME_KEEP = 0,
+    gcvATTRIB_SCHEME_2_10_10_10_REV_TO_FLOAT,
+    gcvATTRIB_SCHEME_BYTE_TO_IVEC4,
+    gcvATTRIB_SCHEME_SHORT_TO_IVEC4,
+    gcvATTRIB_SCHEME_INT_TO_IVEC4,
+    gcvATTRIB_SCHEME_UBYTE_TO_UVEC4,
+    gcvATTRIB_SCHEME_USHORT_TO_UVEC4,
+    gcvATTRIB_SCHEME_UINT_TO_UVEC4,
+    gcvATTRIB_SCHEME_DOUBLE_TO_FLOAT,
+} gceATTRIB_SCHEME;
+
+typedef enum _gceBUFOBJ_TYPE
+{
+    gcvBUFOBJ_TYPE_ARRAY_BUFFER = 1,
+    gcvBUFOBJ_TYPE_ELEMENT_ARRAY_BUFFER  = 2,
+    gcvBUFOBJ_TYPE_GENERIC_BUFFER = 100
+
+} gceBUFOBJ_TYPE;
+
+typedef enum _gceBUFOBJ_USAGE
+{
+    gcvBUFOBJ_USAGE_STREAM_DRAW = 1,
+    gcvBUFOBJ_USAGE_STREAM_READ,
+    gcvBUFOBJ_USAGE_STREAM_COPY,
+    gcvBUFOBJ_USAGE_STATIC_DRAW,
+    gcvBUFOBJ_USAGE_STATIC_READ,
+    gcvBUFOBJ_USAGE_STATIC_COPY,
+    gcvBUFOBJ_USAGE_DYNAMIC_DRAW,
+    gcvBUFOBJ_USAGE_DYNAMIC_READ,
+    gcvBUFOBJ_USAGE_DYNAMIC_COPY,
+
+    /* special patch for optimaize performance,
+    ** no fence and duplicate stream to ensure data correct
+    */
+    gcvBUFOBJ_USAGE_DISABLE_FENCE_DYNAMIC_STREAM = 256
+} gceBUFOBJ_USAGE;
+
+/**
+**  @ingroup gcoVG
+**
+**  @brief  Channel mask values.
+**
+**  This enumeration defines the values for channel mask used in image
+**  filtering.
+*/
+
+/******************************************************************************\
+******************************** VG Enumerations *******************************
+\******************************************************************************/
+
+/**
+**  @ingroup gcoVG
+**
+**  @brief  Tiling mode for painting and imagig.
+**
+**  This enumeration defines the tiling modes supported by the HAL.  This is
+**  in fact a one-to-one mapping of the OpenVG 1.1 tile modes.
+*/
+typedef enum _gceTILE_MODE
+{
+    gcvTILE_FILL,
+    gcvTILE_PAD,
+    gcvTILE_REPEAT,
+    gcvTILE_REFLECT
+}
+gceTILE_MODE;
+
+/******************************************************************************/
+/** @ingroup gcoVG
+**
+**  @brief  The different paint modes.
+**
+**  This enumeration lists the available paint modes.
+*/
+typedef enum _gcePAINT_TYPE
+{
+    /** Solid color. */
+    gcvPAINT_MODE_SOLID,
+
+    /** Linear gradient. */
+    gcvPAINT_MODE_LINEAR,
+
+    /** Radial gradient. */
+    gcvPAINT_MODE_RADIAL,
+
+    /** Pattern. */
+    gcvPAINT_MODE_PATTERN,
+
+    /** Mode count. */
+    gcvPAINT_MODE_COUNT
+}
+gcePAINT_TYPE;
+
+/**
+** @ingroup gcoVG
+**
+**  @brief Types of path data supported by HAL.
+**
+**  This enumeration defines the types of path data supported by the HAL.
+**  This is in fact a one-to-one mapping of the OpenVG 1.1 path types.
+*/
+typedef enum _gcePATHTYPE
+{
+    gcePATHTYPE_UNKNOWN = -1,
+    gcePATHTYPE_INT8,
+    gcePATHTYPE_INT16,
+    gcePATHTYPE_INT32,
+    gcePATHTYPE_FLOAT
+}
+gcePATHTYPE;
+
+/**
+** @ingroup gcoVG
+**
+**  @brief Supported path segment commands.
+**
+**  This enumeration defines the path segment commands supported by the HAL.
+*/
+typedef enum _gceVGCMD
+{
+    gcvVGCMD_END, /*  0: 0x00           */
+    gcvVGCMD_CLOSE, /*  1: 0x01         */
+    gcvVGCMD_MOVE, /*  2: 0x02          */
+    gcvVGCMD_MOVE_REL, /*  3: 0x03      */
+    gcvVGCMD_LINE, /*  4: 0x04          */
+    gcvVGCMD_LINE_REL, /*  5: 0x05      */
+    gcvVGCMD_QUAD, /*  6: 0x06     */
+    gcvVGCMD_QUAD_REL, /*  7: 0x07 */
+    gcvVGCMD_CUBIC, /*  8: 0x08         */
+    gcvVGCMD_CUBIC_REL, /*  9: 0x09     */
+    gcvVGCMD_BREAK, /* 10: 0x0A         */
+    gcvVGCMD_HLINE, /* 11: ******* R E S E R V E D *******/
+    gcvVGCMD_HLINE_REL, /* 12: ******* R E S E R V E D *******/
+    gcvVGCMD_VLINE, /* 13: ******* R E S E R V E D *******/
+    gcvVGCMD_VLINE_REL, /* 14: ******* R E S E R V E D *******/
+    gcvVGCMD_SQUAD, /* 15: ******* R E S E R V E D *******/
+    gcvVGCMD_SQUAD_REL, /* 16: ******* R E S E R V E D *******/
+    gcvVGCMD_SCUBIC, /* 17: ******* R E S E R V E D *******/
+    gcvVGCMD_SCUBIC_REL, /* 18: ******* R E S E R V E D *******/
+    gcvVGCMD_SCCWARC, /* 19: ******* R E S E R V E D *******/
+    gcvVGCMD_SCCWARC_REL, /* 20: ******* R E S E R V E D *******/
+    gcvVGCMD_SCWARC, /* 21: ******* R E S E R V E D *******/
+    gcvVGCMD_SCWARC_REL, /* 22: ******* R E S E R V E D *******/
+    gcvVGCMD_LCCWARC, /* 23: ******* R E S E R V E D *******/
+    gcvVGCMD_LCCWARC_REL, /* 24: ******* R E S E R V E D *******/
+    gcvVGCMD_LCWARC, /* 25: ******* R E S E R V E D *******/
+    gcvVGCMD_LCWARC_REL, /* 26: ******* R E S E R V E D *******/
+
+    /* The width of the command recognized by the hardware on bits. */
+    gcvVGCMD_WIDTH = 5,
+
+    /* Hardware command mask. */
+    gcvVGCMD_MASK = (1 << gcvVGCMD_WIDTH) - 1,
+
+    /* Command modifiers. */
+    gcvVGCMD_H_MOD   = 1 << gcvVGCMD_WIDTH, /* =  32 */
+    gcvVGCMD_V_MOD   = 2 << gcvVGCMD_WIDTH, /* =  64 */
+    gcvVGCMD_S_MOD   = 3 << gcvVGCMD_WIDTH, /* =  96 */
+    gcvVGCMD_ARC_MOD = 4 << gcvVGCMD_WIDTH, /* = 128 */
+
+    /* Emulated LINE commands. */
+    gcvVGCMD_HLINE_EMUL     = gcvVGCMD_H_MOD | gcvVGCMD_LINE, /* =  36 */
+    gcvVGCMD_HLINE_EMUL_REL = gcvVGCMD_H_MOD | gcvVGCMD_LINE_REL, /* =  37 */
+    gcvVGCMD_VLINE_EMUL     = gcvVGCMD_V_MOD | gcvVGCMD_LINE, /* =  68 */
+    gcvVGCMD_VLINE_EMUL_REL = gcvVGCMD_V_MOD | gcvVGCMD_LINE_REL, /* =  69 */
+
+    /* Emulated SMOOTH commands. */
+    gcvVGCMD_SQUAD_EMUL      = gcvVGCMD_S_MOD | gcvVGCMD_QUAD, /* = 102 */
+    gcvVGCMD_SQUAD_EMUL_REL  = gcvVGCMD_S_MOD | gcvVGCMD_QUAD_REL, /* = 103 */
+    gcvVGCMD_SCUBIC_EMUL     = gcvVGCMD_S_MOD | gcvVGCMD_CUBIC, /* = 104 */
+    gcvVGCMD_SCUBIC_EMUL_REL = gcvVGCMD_S_MOD | gcvVGCMD_CUBIC_REL, /* = 105 */
+
+    /* Emulation ARC commands. */
+    gcvVGCMD_ARC_LINE     = gcvVGCMD_ARC_MOD | gcvVGCMD_LINE, /* = 132 */
+    gcvVGCMD_ARC_LINE_REL = gcvVGCMD_ARC_MOD | gcvVGCMD_LINE_REL, /* = 133 */
+    gcvVGCMD_ARC_QUAD     = gcvVGCMD_ARC_MOD | gcvVGCMD_QUAD, /* = 134 */
+    gcvVGCMD_ARC_QUAD_REL = gcvVGCMD_ARC_MOD | gcvVGCMD_QUAD_REL     /* = 135 */
+}
+gceVGCMD;
+typedef enum _gceVGCMD * gceVGCMD_PTR;
+
+/**
+**  @ingroup gcoVG
+**
+**  @brief  Blending modes supported by the HAL.
+**
+**  This enumeration defines the blending modes supported by the HAL.  This is
+**  in fact a one-to-one mapping of the OpenVG 1.1 blending modes.
+*/
+typedef enum _gceVG_BLEND
+{
+    gcvVG_BLEND_SRC,
+    gcvVG_BLEND_SRC_OVER,
+    gcvVG_BLEND_DST_OVER,
+    gcvVG_BLEND_SRC_IN,
+    gcvVG_BLEND_DST_IN,
+    gcvVG_BLEND_MULTIPLY,
+    gcvVG_BLEND_SCREEN,
+    gcvVG_BLEND_DARKEN,
+    gcvVG_BLEND_LIGHTEN,
+    gcvVG_BLEND_ADDITIVE,
+    gcvVG_BLEND_SUBTRACT,
+    gcvVG_BLEND_FILTER
+}
+gceVG_BLEND;
+
+/**
+**  @ingroup gcoVG
+**
+**  @brief  Image modes supported by the HAL.
+**
+**  This enumeration defines the image modes supported by the HAL.  This is
+**  in fact a one-to-one mapping of the OpenVG 1.1 image modes with the addition
+**  of NO IMAGE.
+*/
+typedef enum _gceVG_IMAGE
+{
+    gcvVG_IMAGE_NONE,
+    gcvVG_IMAGE_NORMAL,
+    gcvVG_IMAGE_MULTIPLY,
+    gcvVG_IMAGE_STENCIL,
+    gcvVG_IMAGE_FILTER
+}
+gceVG_IMAGE;
+
+/**
+**  @ingroup gcoVG
+**
+**  @brief  Filter mode patterns and imaging.
+**
+**  This enumeration defines the filter modes supported by the HAL.
+*/
+typedef enum _gceIMAGE_FILTER
+{
+    gcvFILTER_POINT,
+    gcvFILTER_LINEAR,
+    gcvFILTER_BI_LINEAR
+}
+gceIMAGE_FILTER;
+
+/**
+**  @ingroup gcoVG
+**
+**  @brief  Primitive modes supported by the HAL.
+**
+**  This enumeration defines the primitive modes supported by the HAL.
+*/
+typedef enum _gceVG_PRIMITIVE
+{
+    gcvVG_SCANLINE,
+    gcvVG_RECTANGLE,
+    gcvVG_TESSELLATED,
+    gcvVG_TESSELLATED_TILED
+}
+gceVG_PRIMITIVE;
+
+/**
+**  @ingroup gcoVG
+**
+**  @brief  Rendering quality modes supported by the HAL.
+**
+**  This enumeration defines the rendering quality modes supported by the HAL.
+*/
+typedef enum _gceRENDER_QUALITY
+{
+    gcvVG_NONANTIALIASED,
+    gcvVG_2X2_MSAA,
+    gcvVG_2X4_MSAA,
+    gcvVG_4X4_MSAA
+}
+gceRENDER_QUALITY;
+
+/**
+**  @ingroup gcoVG
+**
+**  @brief  Fill rules supported by the HAL.
+**
+**  This enumeration defines the fill rules supported by the HAL.
+*/
+typedef enum _gceFILL_RULE
+{
+    gcvVG_EVEN_ODD,
+    gcvVG_NON_ZERO
+}
+gceFILL_RULE;
+
+/**
+**  @ingroup gcoVG
+**
+**  @brief  Cap styles supported by the HAL.
+**
+**  This enumeration defines the cap styles supported by the HAL.
+*/
+typedef enum _gceCAP_STYLE
+{
+    gcvCAP_BUTT,
+    gcvCAP_ROUND,
+    gcvCAP_SQUARE
+}
+gceCAP_STYLE;
+
+/**
+**  @ingroup gcoVG
+**
+**  @brief  Join styles supported by the HAL.
+**
+**  This enumeration defines the join styles supported by the HAL.
+*/
+typedef enum _gceJOIN_STYLE
+{
+    gcvJOIN_MITER,
+    gcvJOIN_ROUND,
+    gcvJOIN_BEVEL
+}
+gceJOIN_STYLE;
+
+/* Base values for channel mask definitions. */
+#define gcvCHANNEL_X    (0)
+#define gcvCHANNEL_R    (1 << 0)
+#define gcvCHANNEL_G    (1 << 1)
+#define gcvCHANNEL_B    (1 << 2)
+#define gcvCHANNEL_A    (1 << 3)
+
+typedef enum _gceCHANNEL
+{
+    gcvCHANNEL_XXXX = (gcvCHANNEL_X | gcvCHANNEL_X | gcvCHANNEL_X | gcvCHANNEL_X),
+    gcvCHANNEL_XXXA = (gcvCHANNEL_X | gcvCHANNEL_X | gcvCHANNEL_X | gcvCHANNEL_A),
+    gcvCHANNEL_XXBX = (gcvCHANNEL_X | gcvCHANNEL_X | gcvCHANNEL_B | gcvCHANNEL_X),
+    gcvCHANNEL_XXBA = (gcvCHANNEL_X | gcvCHANNEL_X | gcvCHANNEL_B | gcvCHANNEL_A),
+
+    gcvCHANNEL_XGXX = (gcvCHANNEL_X | gcvCHANNEL_G | gcvCHANNEL_X | gcvCHANNEL_X),
+    gcvCHANNEL_XGXA = (gcvCHANNEL_X | gcvCHANNEL_G | gcvCHANNEL_X | gcvCHANNEL_A),
+    gcvCHANNEL_XGBX = (gcvCHANNEL_X | gcvCHANNEL_G | gcvCHANNEL_B | gcvCHANNEL_X),
+    gcvCHANNEL_XGBA = (gcvCHANNEL_X | gcvCHANNEL_G | gcvCHANNEL_B | gcvCHANNEL_A),
+
+    gcvCHANNEL_RXXX = (gcvCHANNEL_R | gcvCHANNEL_X | gcvCHANNEL_X | gcvCHANNEL_X),
+    gcvCHANNEL_RXXA = (gcvCHANNEL_R | gcvCHANNEL_X | gcvCHANNEL_X | gcvCHANNEL_A),
+    gcvCHANNEL_RXBX = (gcvCHANNEL_R | gcvCHANNEL_X | gcvCHANNEL_B | gcvCHANNEL_X),
+    gcvCHANNEL_RXBA = (gcvCHANNEL_R | gcvCHANNEL_X | gcvCHANNEL_B | gcvCHANNEL_A),
+
+    gcvCHANNEL_RGXX = (gcvCHANNEL_R | gcvCHANNEL_G | gcvCHANNEL_X | gcvCHANNEL_X),
+    gcvCHANNEL_RGXA = (gcvCHANNEL_R | gcvCHANNEL_G | gcvCHANNEL_X | gcvCHANNEL_A),
+    gcvCHANNEL_RGBX = (gcvCHANNEL_R | gcvCHANNEL_G | gcvCHANNEL_B | gcvCHANNEL_X),
+    gcvCHANNEL_RGBA = (gcvCHANNEL_R | gcvCHANNEL_G | gcvCHANNEL_B | gcvCHANNEL_A),
+}
+gceCHANNEL;
+
+/* Defines the statistical data keys monitored by the statistics module */
+typedef enum _gceSTATISTICS
+{
+    gcvFRAME_FPS        =   1,
+}
+gceSTATISTICS;
+
+/* Value types. */
+typedef enum _gceVALUE_TYPE
+{
+    gcvVALUE_UINT = 0x0,
+    gcvVALUE_FIXED,
+    gcvVALUE_FLOAT,
+    gcvVALUE_INT,
+
+    /*
+    ** The value need be unsigned denormalized. clamp (0.0-1.0) should be done first.
+    */
+    gcvVALUE_FLAG_UNSIGNED_DENORM = 0x00010000,
+
+    /*
+    ** The value need be signed denormalized. clamp (-1.0-1.0) should be done first.
+    */
+    gcvVALUE_FLAG_SIGNED_DENORM   = 0x00020000,
+
+    /*
+    ** The value need to gammar
+    */
+    gcvVALUE_FLAG_GAMMAR          = 0x00040000,
+
+    /*
+    ** The value need to convert from float to float16
+    */
+    gcvVALUE_FLAG_FLOAT_TO_FLOAT16 = 0x0080000,
+
+    /*
+    ** Mask for flag field.
+    */
+    gcvVALUE_FLAG_MASK            = 0xFFFF0000,
+}
+gceVALUE_TYPE;
+
+typedef enum _gceTRACEMODE
+{
+    gcvTRACEMODE_NONE     = 0,
+    gcvTRACEMODE_FULL     = 1,
+    gcvTRACEMODE_LOGGER   = 2,
+    gcvTRACEMODE_ALLZONE  = 3,
+    gcvTRACEMODE_PRE      = 4,
+    gcvTRACEMODE_POST     = 5,
+} gceTRACEMODE;
+
+enum
+{
+    /* GPU can't issue more that 32bit physical address */
+    gcvPLATFORM_FLAG_LIMIT_4G_ADDRESS = 1 << 0,
+
+    gcvPLATFORM_FLAG_IMX_MM           = 1 << 1,
+};
 
 /* No special needs. */
 #define gcvALLOC_FLAG_NONE                  0x00000000
@@ -2392,7 +2137,6 @@ gceVIDMEM_TYPE;
 /******************************************************************************\
 ****************************** Object Declarations *****************************
 \******************************************************************************/
-
 typedef struct _gckCONTEXT          * gckCONTEXT;
 typedef struct _gcoCMDBUF           * gcoCMDBUF;
 
@@ -2400,8 +2144,10 @@ typedef struct _gcsSTATE_DELTA      * gcsSTATE_DELTA_PTR;
 typedef struct _gcsQUEUE            * gcsQUEUE_PTR;
 typedef struct _gcoQUEUE            * gcoQUEUE;
 typedef struct _gcsHAL_INTERFACE    * gcsHAL_INTERFACE_PTR;
+#if VIVANTE_PROFILER
+typedef struct _gcsHAL_PROFILER_INTERFACE    * gcsHAL_PROFILER_INTERFACE_PTR;
+#endif
 typedef struct _gcs2D_PROFILE       * gcs2D_PROFILE_PTR;
-
 
 #if gcdENABLE_VG
 typedef struct _gcoVGHARDWARE *            gcoVGHARDWARE;
