+++ kernel-module-imx-gpu-viv-src/hal/kernel/inc/gc_hal_metadata.h
@@ -66,33 +66,6 @@ extern "C" {
 
 #define VIV_VIDMEM_METADATA_MAGIC __FOURCC('v', 'i', 'v', 'm')
 
-/* Compressed format now was defined same as dec400d, should be general. */
-typedef enum _VIV_COMPRESS_FMT
-{
-    _VIV_CFMT_ARGB8 = 0,
-    _VIV_CFMT_XRGB8,
-    _VIV_CFMT_AYUV,
-    _VIV_CFMT_UYVY,
-    _VIV_CFMT_YUY2,
-    _VIV_CFMT_YUV_ONLY,
-    _VIV_CFMT_UV_MIX,
-    _VIV_CFMT_ARGB4,
-    _VIV_CFMT_XRGB4,
-    _VIV_CFMT_A1R5G5B5,
-    _VIV_CFMT_X1R5G5B5,
-    _VIV_CFMT_R5G6B5,
-    _VIV_CFMT_Z24S8,
-    _VIV_CFMT_Z24,
-    _VIV_CFMT_Z16,
-    _VIV_CFMT_A2R10G10B10,
-    _VIV_CFMT_BAYER,
-    _VIV_CFMT_SIGNED_BAYER,
-    _VIV_CFMT_VAA16,
-    _VIV_CFMT_S8,
-
-    _VIV_CFMT_MAX,
-} _VIV_COMPRESS_FMT;
-
 /* Metadata for cross-device fd share with additional (ts) info. */
 typedef struct _VIV_VIDMEM_METADATA
 {
