+++ kernel-module-imx-gpu-viv-src/hal/os/linux/kernel/gc_hal_kernel_os.c
@@ -2271,7 +2271,11 @@ gckOS_MapPhysical(
         {
             /* Map memory as cached memory. */
             request_mem_region(physical, Bytes, "MapRegion");
+#if LINUX_VERSION_CODE >= KERNEL_VERSION(5,5,0)
+            logical = (gctPOINTER) ioremap(physical, Bytes);
+#else
             logical = (gctPOINTER) ioremap_nocache(physical, Bytes);
+#endif
 
             if (logical == gcvNULL)
             {
@@ -5190,9 +5194,15 @@ gckOS_GetProfileTick(
     OUT gctUINT64_PTR Tick
     )
 {
+#if LINUX_VERSION_CODE >= KERNEL_VERSION(5,5,0)
+    struct timespec64 time;
+
+    ktime_get_ts64(&time);
+#else
     struct timespec time;
 
     ktime_get_ts(&time);
+#endif
 
     *Tick = time.tv_nsec + time.tv_sec * 1000000000ULL;
 
@@ -5204,7 +5214,11 @@ gckOS_QueryProfileTickRate(
     OUT gctUINT64_PTR TickRate
     )
 {
+#if LINUX_VERSION_CODE >= KERNEL_VERSION(5,5,0)
+    struct timespec64 res;
+#else
     struct timespec res;
+#endif
 
 #if LINUX_VERSION_CODE >= KERNEL_VERSION(4, 2, 0)
     res.tv_sec = 0;
